make transformations post-hoc
