+++ src/lexicon/_lexicon.js
@@ -1,1 +1,1 @@
-module.exports={"Noun":"0:2K;1:1V;a27b1Yc14d0Ue0Nf0Mg0Jh0Di05journ1lZmQnews servi2EoOpIquartet,rFs6t4u2vs,worke1F;ni2tilD;on,v7;ce,ele2im2Hribun1v;communica1Egraph,v0S;av0Achool,e8o5t2ubcommit1Fyndicat;ate 2ockP;poli27univ2;ersi23;ci2ns;al club,et2;e,y;cur2rvice0;iti27;adio,e2;gionLs2;ourc24tauraS;artners6d,e5harmac4izza,l3ol2roduc11;i1Wytechn0A;!c;a0Geutical0;ople's par1Qtroleum;!hip;bservHffi1Rptic1r2;chestra,ganiz1Z;a7d,e5i3otor0use2;e,um;lita1Bnist2;e08ry;dia,mori1rcantile2; exchange;!ch1Ngazi3nage06r2;i2ket0D;ne;ab5i4oc2;al 2;aIheaH;beration army,t;or2s;ato0Z;e,n2;dustri1Fs5ter4vest2;me2o09;nt0;nation1sta11;titut2u14;!e2;! of technology,s;e4o2;ld2sp0Itel0;ings;a2ra5;lth a2;uth0T;a3ir0AovernJroup,ui2;ld;s,zet0P;acul0QedRm,ound12und,y;g7lectr6n5s4t3veningHx2;!p;at,c;p,ta0K;semb01terprise0;ic;!li05;a9e4i2rillG;rectora0Fv2;is0T;part2velop5;e4ment2;! sto2s;re;ment;ily2ta; news;aRca,entrQhNircus,lLo2rew;alit0KlHm9n3rp2unc6;!or0Is;fe5s2taine9;e3ult2;ing;il;d2ren04;er0C;m4p2;any,rehensAute2;rs;i4uni2;ca2ty;tions;ss06t2;tee;!le2;ct2ge;ive;ini2ub;c,qI;emica3oir,ronic2urch;le;ls;al bank,e;!fe,is4p2thedr1;it1;al;se;a7c,l,o6r3u2;ilding socieDreau;ands,ewe3other2;hood,s;ry;ard,ys;!n2;k,q2;ue;dGgencDirBrAss6ut2viH;h3ori2;te;ori2;ty;oc4u2;ran2;ce;!i9;c,my;l2ways;in3;e0i2y;es;!s;!ministr2;at2;ion","Uncountable":"0:1E;a1Jb1Dc14e0Yf0Sg0Oh0Ii0Fj0Eknowled1Il09m00nZoYpSrPsCt8vi7w1;a5ea0Ci4o1;o2rld1;! seJ;d,l;ldlife,ne;rmth,t0;neg7ol0A;e3hund0ime,oothpaste,r1una;affTou1;ble,sers,t;a,nnis;aBceWeAh9il8now,o7p4te3u1;g1nshi0N;ar;am,el;ace2e1;ciPed;!c13;ap,cc0ft0C;k,v0;eep,opp0P;riK;d08fe0Xl1nd;m0Rt;ain,e1i0X;c1laxa0Dsearch;ogni0Crea0C;a5e3hys0FlastAo2r1;ess00ogre03;rk,w0;a1pp0trol;ce,nR;p0tiK;il,xygen;ews,oi0C;a7ea5i4o3u1;mps,s1;ic;nHo08;lk,st;sl1t;es;chi1il,themat00;neD;aught0e3i2u1;ck,g07;ghtnZqu0CteratI;a1isH;th0;ewel7usti08;ce,mp1nformaOtself;ati1ortan06;en05;a4isto3o1;ck1mework,n1spitali01;ey;ry;ir,lib1ppi9;ut;o2r1um,ymnastJ;a7ound;l1ssip;d,f;i5lour,o2ruit,urnit1;ure;od,rgive1wl;ne1;ss;c6sh;conom9duca5lectriciMn3quip4th9very1;body,o1thB;ne;joy1tertain1;ment;tiC;a8elcius,h4iv3loth6o1urrency;al,ffee,n1ttA;duct,fusi9;ics;aos,e1;e2w1;ing;se;ke,sh;a3eef,is2lood,read,utt0;er;on;g1ss;ga1;ge;c4dvi3irc2mnes1rt;ty;raft;ce;id","PresentTense":"0:7F;1:7G;2:7H;3:73;4:67;5:5X;6:7A;7:5P;8:47;a6Ob66c56d4De3Vf3Kg3Dh37i2Vj2Tk2Ql2Hm26n23o1Zp1Equest46r0RsYtOuKvHwAy9;aw2iel6;aCe1Zhistl0iBo9rec4;n25r9;ks,ri0;ns,p0sh0t68;l4nAr9sh0t6Wv0;ns,ran1;de3ts;aAi9ot0;e73si1;lu0ri0;n9plif1rg0s0;d9it0l61;er9os;g3Xm5D;aHeGhEoDrAu9w1H;n0r2;aAea1i9us1;c4es;d0i2ns48ps,ve5;o4rt3Mu6H;an4r9;eate2iv0;n6s1;k0l4r5Kst0;a00cXeThQid0ki8lPmOnMor1pKtDu9wit6C;bmi1c4f41gg0TppAr9sp40;g0pr38roun6v66;li0o9;r1s0;aEeDoCrAu9;di0m60;ai2e9iv0;ngthe2ss0t64;ps,r0;ms,ps;i2k0r0Qys;i1la59oi5rin9;gs,kl0;e9o69;a4ez0;a55el5il0ok0;a7ee8i8;aAo9;c4u1;r0v0;ar5ReBl3GnArv0t9;s,tl0;ds,s0;ks,ms;aAor0r9;at5Me5X;l0r0;i5v0;aTeBis4oAu9;bs,i2l0sh0;a7l5;aPcOdZfNgMj35lJmIn1pGquFsCtBv9war6;e9ie5Q;a5ls,rs0;ai2ir0ur2;eAis1or1t9;rYs;m54n1rv0;es1ir0;ai3l9or1resU;ac0i0;ai2embe3in6ov0;aAen1i9;es,v0;t0xs;ar6re1;l2Qus0;al5e4Wko2ogn29;ch0l28so2;is0n9;g0ks;aSerQiPlMoKrBu9;b1Wl5m8r9sh0ts;ch2Usu0;a1ReEin1o9;cCdBgre57hi0Tm1QtAv9;es,id0;es1;uc0;e53lai7;dApar0ss0v9;en1;ic1;in1ps,rt3Zs9u3;es,ts;aAu9;gs,ng0s0;ns,ys;c4l0;for7mi1s9;is1;c4in1rt4Gss0us0ys;bBccu3f21per4TutweAver9w0;c3Tla8tak0us0whel7;ig2J;se22t2W;eAot9;es,ic0;e6gl1U;aFeEiCoBu9;ltipli0r9;de3;nito3v0;n9rro3;ds,gl0us0;as13l1nt12rg0ss21;int2Lke du0n20rAt9;ch0te3;k9ri0;e1s;aFeEiBo9um8;a9c4o4v0;ds,ns;f1k0mi1nAst9v0;e2s;es,ks;ar2ct0S;bo3c4u9;g1Vn3L;ee8iAno9;c4ws;c4l5ss0;a7oi2u9;dg0m8;gnor0mIn9ssu0;cGdic3Wflue3KhEsCterBv9;e9olv0;n1s1;ac1fer0;is1t9;al5r1T;a9i9;bi1;lud0re16;ag1Vp9;or1r0S;aDeBir0oAu9;n1rri0;ld fMno3p0s1;a9l8;ds,ts;ndl0ppe2r7t0;aEe1lo3CoDrBu9;ar9e3Bid0;ante0ds;a9i8o35um2O;bs,s8;es,v1D;i2t22z0;aIeaHiFlDo9ractVunctU;c2WlBr9;c0ec9ms,se0;as1;ds,lo2X;a1Uo9;a1o6ws;gOl9ni1Sxs;es,ls;rs,tM;i5l5vo3;ar2chPdi1l04mMnHstFvalu2Xx9;am14cDerBis1p9ten6;an6e9l11or1re2S;c1rie2I;c9ts;is0;e5ha18;ab9im2Q;li1G;a20cour09dCg09ha2Cjo1Mri26sBte3vi9;es,s9;io2;ur0;s,ur0;bodi0erg0p9;has9o25;iz0;os;aZeOiEoDr9um8;aBeAif1o8;ps;a7ss0;gs,i2;dg0es,ub1;es,fHps,rGsAv9;es,id0orc0;aBc9lik0mi24pla15tingui0Y;o9u23;urQve3;g1Ypp9;ea3r9;ov0;ec1;fe3;cGe7fDlive3m06ni0pCs9term07;crib0eAi9tro0W;g2r0;rv0;ar1en6ic1osi1;eAi9;es,n0;a1c1;id0lAre9;as0;ar0in0;m9nc0r0t0;ag0;a05e04h00it0lWoCr9u1;aAe9i0o1Iu0C;at0di1;c4v0;llRmMnBo4pi0u9ve3;g9n1;hs;cIdHfGsDtBv9;e9i12;r1ys;ai2inu0r9;as1ibAo5;iAtit9;ut0;de3s1;ir7or7ron1;uc1;er2;bCes,for1mBp9;ar0l9;ai2;an6;in0;aps0ec1;aAi9os0;nXps;i7ws;ms;aAe9;a1c4ws;lle9ng0rg0;ng0;as0nte3;l5nce5r9te3us0;es,ri0;ls;aOeIlGoErCu9;dAr9;i0ns;ge1;ak0u9;sh0;as1mbs,os1rro09t9u04;he3;oc4;ks;a3cDlBnefi1t9;ra9;ys;i0Aon9;gs;om0;n9s0;gs,s;bando2c03dZffor6gYi6lUmTnPppKrHsCtt9voi6wai1;acAe9rac1;mp1n6;h0ks;ks,s9;eAi9ociZum0;g2s1;m9r1;bl0;chAgu0r9;iv0;es,iv0;eaClBro9;a9v0;ch0;au6i0;ls,rs;nouBsAt9;icipM;we3;nc0;us0;lAte3;rs;eg0o9;ws;re0;d9jus1mi1op1;e6re9s;ss0;ds;cBhiAknowledg0tiv9;at0;ev0;ep1oun1us0;es;ts;ns","Infinitive":"0:6A;1:6K;2:6N;3:6F;4:6L;5:5P;6:4V;7:6J;8:5X;9:68;a5Wb5Ic4Md3Ve3Gf33g2Yh2Ri2Gj2Ek2Bl22m1Un1Ro1Np18ques34r0MsXtOuKvIwByA;awn,ield;aDe1Nhist8iCoAre5N;nd0rA;k,ry;n,pe,sh,th0;lk,nBrAsh,t66ve;n,ra6Q;d0t;aAi0Io7;l5Yry;nApli3Nr1se;dAi7lo5D;erAo;go,mi4T;aHeGhEoDrBuAwi3;ne,rn;aAi58u3y;de,in,nsf0v4Z;ok,r2Mu5T;ank,rA;eat2Ci4;nd,st;ke,lk,s7;aXcUeRhPi3WlOmNneMo4IpKtEuAwit5O;ck,ff0gge3ppBrAspe5;ge,pri2rou4Fvi4;ly,oA;rt,se;aEem,oDrBuA;dy,mb8;a48eAi4;ngth21ss,t5G;p,re;in,ke,r0Jy;la4Moil,rinA;g,k8;ak,e2T;a4Je1Li8o4K;am,e1A;aAo4L;re,ve;ar56eBle5nAr4tt8;d,se;k,m;aBo6rA;at52ew;le,re;il,ve;aTeCisk,oBuA;b,in,le,sh;am,ll;aPcOdu9fNga56je5lLmKnt,pIquHsEtDvAwa56;eBiA;ew;al,l,r2;a3Ji6u3G;eBi3o3LtA;!ri5;mb8nt,r4;e3i6;air,lAo3Hrese58;a9y;a3Cemb0i3Eo4;aAe55i4y;te,x;le5u2;a0Sei4k54ogni1Z;ch,li1Ys53;i2nA;ge,k;aNerMiLlJoHrCuA;b1Oll,mp,rAsh;cha2s47;ai2eDi4VoA;cBdu9mi2te3vA;e,i2G;e4Llaim;di5pa6ss,ve4R;i4Qp,rtr3KsAur;e,t;a3AuA;g,n1;ck,le;fo2Nsi3;ck,i4Krt43ss,u2y;bCccur,ff0pera7utweBverAwe;co3Ota37u2whelm;igh;ser4ta2M;eBotA;e,i9;ed,gle5;aFeEiCoBuA;ltiply,rd0;nit0Sve;nArr0R;d,g8us;asu6lt,n0Er1ssa1;inta2Cke d3Hna1r02tA;ch,t0;aGeFiCoAu13;aAck,ok,ve;d,n;ft,ke,nBstAve;!en;e,k;a21c06;b0Fck,uA;gh,n3B;eBiAno2A;ck,ll,ss;ep;am,o1YuA;d1mp;gno6mInAss31;cGdica7flu0CsDterCvA;eAol4;nt,st;a5fe6;i3tA;aAru5;ll;lu15r13;agi1RpA;o1Pro4;aEeDi6oCuA;nt,rA;ry;ld fa3nUpe,st;ad,lp;nd8ppArm,te;en;aDet,lo31ove1BrBuA;arant2Ve30i0U;aAow,umb8;b,sp;in,th0ze;aKeaJiHlFoCracBuncA;ti34;tu6;cus,lBrA;ce,eca3m,s2N;d,l1P;a1LoA;od,w;gu6lAni1Jx;e,l;r,tu6;il,ll,vA;or;a0Tcho,le5mLnIstGvalua7xA;a03cEerDi3pAte0X;a0WeAla0Uo0Xre2H;ct,riA;en9;ci2t;el,han1;abAima7;li16;ab8couRdBga1han9jXri1Wsu6t0vA;isi2Iy;!u6;body,er1pA;hasiAow0;ze;aZeOiGoFrBuA;mp;aCeBiAop;ft;am,ss;g,in;!d1ubt;e,ff0re5sBvA;e,iQor9;aDcAli0Rmi1Vpl0Wtingui0Q;oAu1U;uAv0;ra1;gr1MppA;ear,ro4;cGem,fFliv0ma04ny,pEsBterA;mi05;cribe,er4iBtrA;oy;gn,re;a00eZi5;e5i00y;iClBrA;ea2;a6iX;de;ma1n9re,te;a02e01hYi7lWoCrA;aAea7o1Cu07y;ck,ve;llSmLnBok,py,uAv0;gh,nt;ceIfGsEtCvA;eAin9;rt,y;aIin0NrA;a3ibu7ol;iAtitu7;d0st;iAoAro18;rm;rn;biFfoEmaDpA;a6laA;in;re;nd;rt;ne;ap2e5;aAiTo2;im,w;aBeA;ck,w;llen1n1r1;a2nt0;ll,ncBrAt0u2;e,ry;el;aLeHloGoErBurA;n,y;aBuA;sh;ke;a3mb,o3rrAth0un9;ow;ck;ar,lBtrA;ay;ie4oA;ng;nAse;!g;band0Cc04d00ffoZgrYid,lVmu2nSppNrKsEttAvoid;acCeBra5;ct;m08nd;h,k;k,sA;eCiBocia7uA;me;gn,st;mb8rt;le;chBgAri4;ue;!i4;eaDlCroA;aAve;ch;aud,y;l,r;noun9sw0tA;icipa7;ce;lBt0;er;e1ow;ee;rd;dBju3oK;st;!ed,reA;ss;cDhie4knowled1tiva7;te;ge;ve;eCouBu2;se;nt;pt;on","Organization":"0:43;1:3K;a38b2Pc29d20e1Xf1Tg1Lh1Gi1Dj19k17l13m0Tn0Go0Dp07qu06rZsStFuBv8w3y2;amaha,m0Youtu2Rw0Y;a4e2orld trade organizati40;lls fargo,st2;fie22inghou17;l2rner br3C;-m12gree30l street journ25m12;an halOeriz3Visa,o2;dafo2Gl2;kswagMvo;bs,n3ps,s2;a tod2Qps;es34i2;lev2Wted natio2Tv; mobi2Jaco beQd bNeBgi fridaAh4im horto2Smz,o2witt2V;shiba,y2;ota,s r Z;e 2in lizzy;b4carpen32daily ma2Wguess w3holli0rolling st1Ms2w3;mashing pumpki2Nuprem0;ho;ea2lack eyed pe3Eyrds;ch bo2tl0;ys;l3s2;co,la m13;efoni09us;a7e5o3pice gir6ta2ubaru;rbucks,to2M;ny,undgard2;en;a1x pisto2;ls;few24insbu25msu1W;.e.m.,adiohead,b7e4oyal 2yan2W;b2dutch she5;ank;/max,aders dige1Ed 2vl31;bu2c1Thot chili peppe1lobst27;ll;c,s;ant2Uizno2E;an6bs,e4fiz23hilip morrCi3r2;emier26octer & gamb1Qudenti14;nk floyd,zza hut;psi27tro2uge09;br2Pchina,n2P; 3ason1Wda2F;ld navy,pec,range juli3xf2;am;us;aBbAe6fl,h5i4o2sa,wa;kia,tre dame,vart2;is;ke,ntendo,ss0K;l,s;stl4tflix,w2; 2sweek;kids on the block,york09;e,é;a,c;nd1Ss3t2;ional aca2Do,we0P;a,cYd0N;aAcdonald9e6i4lb,o2tv,yspace;b1Lnsanto,ody blu0t2;ley crue,oro0N;crosoft,t2;as,subisO;dica3rcedes-benz,talli2;ca;id,re;'s,s;c's milk,tt12z1X;'ore09a4e2g,ittle caesa1td;novo,x2;is,mark; pres6-z-boy;atv,fc,kk,m2od1J;art;eopardy,iffy lu0Ko4pmorgan2sa;! cha2;se;hnson & johns1Rom06y d1Q;bm,hop,n2tv;c,g,te2;l,rpol; & m,asbro,ewlett-packaTi4o2sbc,yundai;me dep2n1I;ot;tac2zbollah;hi;eneral 7hq,l6o3reen d0Hu2;cci,ns n ros0;ldman sachs,o2;dye2g0A;ar;axo smith kliZencore;electr0Hm2;oto1;a4bi,da,edex,i2leetwood mac,oGrito-l09;at,nancial2restoV; tim0;cebook,nnie mae;b05sa,u,xxon2; m2m2;ob0G;aiml0Ae6isney,o4u2;nkin donuts,po0Wran dur2;an;j,w j2;on0;a,f leppa4ll,p3r spiegZstiny's chi2;ld;eche mode,t;rd;aFbc,hCiAnn,o4r2;aigsli6eedence clearwater reviv2;al;!ca co6l5m2o1st05;ca3p2;aq;st;dplMgate;la;a,sco2tigroup;! systems;ev3i2;ck fil-a,na daily;r0Hy;dbury,pital o2rl's jr;ne;aGbc,eCfAl6mw,ni,o2p,rT;ei4mbardiKston 2;glo2pizza;be;ng;ack & deckGo3ue c2;roY;ckbuster video,omingda2;le; g2g2;oodriO;cht4e ge0n & jer3rkshire hathaw2;ay;ryI;el;nana republ4s2xt6y6;f,kin robbi2;ns;ic;bYcTdidSerosmith,ig,lMmGnheuser-busFol,ppleBr8s4t&t,v3y2;er;is,on;hland3s2;n,ociated G; o2;il;by5g3m2;co;os; compu3bee2;'s;te1;rs;ch;c,d,erican4t2;!r2;ak; ex2;pre2;ss; 5catel3t2;air;!-luce2;nt;jazeera,qae2;da;as;/dc,a4er,t2;ivisi2;on;demy of scienc0;es;ba,c","Singular":"ad hominSbOcMdIeGfEgDhAkitNlunch,n9others,p7room,s5t0us dollar,weekFyesterd2;h2ic,o0ragedy;d0morrow,night;ay;ere,ing0;!s;ky,tu0uper bowl,ystM;dAff;a0roblKurpo4;rt,t8;othHumbB;ead start,o0;me,u0;se;laci7od,rand slam;ri0undC;end;conomy,gg,ner3v0xamp9;ent;eath,inn2o0;g0or;gy;er;anary,eil4i0ottage,redit card;ty;ank2o0reakfast;dy,tt0;le;ing;em","Adjective":"0:JR;1:JS;2:IT;3:FI;4:JP;5:JI;6:JJ;7:J5;8:IB;9:JQ;A:EY;B:GO;C:GL;D:J7;E:J8;F:IW;aHHbGBcEEdCSeBMfAHg9Vh95i87j83k7Xl7Am6Kn69o5Pp4Pqu4Mr3Ns1Kt0Yu09v03wIyGza5D;awn0eGield0ou5;ar3ll9C;aWeShPiKoHrG;e0Jo5y;man3nd8oHrGuH4;k0ri0;dH7zy;ck2deJlInHpe0ry,sGth8;e,h0tfH2;dy,n0;d,y;!sprIJ;iHoleG;!sa1;stHMte;aIek3iHlGsteCAt;coH4l oB1;gh0rd;k,ry;lk0nIrHsGtHTve0;h0t9K;m,n0p2raJ4y;d8t0;aJeIiHoGulgar;laAIte0;bJDce versa,ew0olCNsIK;ng9ErJ2xi5;ca4gGin,lHEnilla,ri0st;aboGGue;g3lt03nMpJrIsHtterG;!moI5;eFBu7;bD9ge0; f7Xb56coHWliBYpHsGt4W;ca1et,ide dPtairs;er,iAF;aUbecoHTconvin5WdQePfair,iOknNlLprecede0IrJsHtiGCwG;ielGBrittGC;ig5CuG;pervis2spec7D;eGu3;cognQgulHDlHD;ike3oG;ck0;own;te0vers7;ndi5vG3xpect2;eGo0;cid2rG;!gGmiE3stood;o0rouFU;ppeaZssuHEuthorG;iz2;i6Wra;aXeUhRiQoNrHuGwiF;ne0rn0;aJeIiHoubl2uG;e,st0;ab1ck0te;at0nFP;de0i9nGp0vEJ;quE4sf8;gethBok0p,rHuG;ch0gh;pH9tuA;dy,gEQme3ny,r2;ank0iHoughBYrG;eate9iE;ck,n,rs9I;enaFQlHmD1nGrr7Dst9Z;d0se;li5;boo,ke0lHme,n,rGsteE6wdGL;gECt;eGk0l;nt2;a1Fc1Ae13h0Zi0Xki0Wl0Rm0On0Lo0Gp09qu08tRuIwGymbGN;eB7iG;ft,tG6;bMck0ddF6ff8ggeFi generEGpIrGspeCY;e,ge0prisA4rGviE;e7ouD5;erIpGreFA;li0oG;rt0se0;! dupBb,fiFPi6C;du2mGVsGt1;eqCNtan9R;aTeQiPoOrHuGyl8D;di0mbFJpGKrEV;aIeHiGo5;ct,p2ve0;ngthe9ss0tFS;iGnF2;ghtGnn0;! fGfG;orEU;p0re0;ff,ll;adHep,m0p0rG;eotyp2n;faG6y;iHke0le,ndaH6rG;e0k,t0;ab1nn0;aCeam7W;aCeciLiKlJoIriG;g3FnG;g0kF0;il0t on;aDDendFZ;cy,rit2t0;al,fFO; call2ci7ft,lImbBotEQphist57rHuGviA7;r,th9I;dFUe,ry,t0;eGid;!mn;eHoG;bb7Iw0;ak0e94;aHell68iENoGug;ke0o3R;ll,rt,sh0;am0eIiGow,y;ck,gD2mGp0;!y;eGndB;k,p0;n1Rp0;ck,de0gnifica4l3mp1nG;ceCg1;aHeBi1OoGri26ut,y;ck49rt,ut0;dy,ll5Nm6ErGve0;e0p;arEKcLda6eKlJnIpaA1rHtGveCxy;ab1tE7;e7Eve0;d0i4Use0;eB9f6X;k0m3P;o3Sr9EuC;aIient5BoArG;atECeG;ecDUw0;le0rGtDT;ce,e0y;cr2d,fe,iHl7Ame,ns serif,vG;e0vy;l0nt3;a0BeLiJoHuG;b0de,i9le0sh0;am0buEQll0maDGo6Msy,ttD4uGy7;gh,nd,ti6Y;ch,ghtGpe,sk0tzy,v7;!-5CfCY;a03bel,c01dZfYgXjeASlTmRnt0pPquOsKtIvGwaEP;eGiew0;al0l0rB5;aGiAuAR;i9rd2;eIista9EoHtG;ab1riAM;lu6na4rt0;mbDFnt0rE;eFiA;air0lGoASreseF2ublic9I;aDWi0;ai9emb8iAPoG;te,ve0;aIeHiGuc6X;ab1ve0;nt0va4;te0x0;aE7rBI;leA9resCZuDR;!uG;ce0nEU;all0eiEko9oG;gni7Gndi6;ch0dy,lGr,so9;!i7E;bDYciHgg2iDKnGpDYre,w;dBMge0k0;al,st;aHest73iG;ck,et;ck,i4;a08e04ho03i01lXoUrJuG;bliHll0mp6Mny,rGsh0t0;chaDCe,p1sCU;c,sh0;aiDAeOiLoG;!b0cJduD6fIgreE0hibDVmiHpBteFud,vG;e0i8N;ne4se0;ouBRuDO;eDWlaim0;cHmGnt0or,sti5Pva6;aD1e;ey,k3;ciDJdi9Jg7FmGpaAss0veE3;a15iDT;iE1li6or,pHrtrBFsGt6Hur0;e0t0;ab1uliD9;aHeasa4uG;g0nDs0;cD8i8TnG;n0t;cGle0nk;ayu5Ck0;ny;ac46rGti6;fHmGplexi5siFverD6;a6Hit0;eDTo93;ck0iKle,rHsGuCL;s0t;allel2tG;-tiBFiG;al,cCD;ab1nG;fB2t0;bVccur0dd,ffTi3kSldRnQpOrNthButMvGwe0;al,erG;!aJcoBMlap0nItHuCCwG;eHhelm0rougA7;!a6X;igA5;ll;do29er,g67siB8weigh0;anB6g2Y;eGposi6;n,raD1; boaDGg63li4O;! fashion2;!ay; guaDDbGer0iBFli4L;eat;eIsGt54;cGerv2Dole6;e4HuC;d9Qse;aPeMiKoGuaCD;b1isy,nIrHsy,tG;ab1e0iBP;m7th5S;chala4descriD1stop;ce,gGmb1;ht3;aGcessaBLed0gle85ighbor3w,xt;rGt;!by;iBVk2rr2A;aWePiMoIuG;ltiplGnda43rd8te;e,i0;bi1derIist,lA5nHoGst,v0;dy,t;it0Mth3;a6n;ld,nGrr0K;d0gAOiaGor,us0;tuC;aKdiC2ek,lJnHrGssaD;e,ge0;aGt4T;ci5;l1Vt0;n,sGty;ly,uA;cNd,genta,inMj14keLle,mmoKnJrHsculi3NtG;ch0t8uC;kGri0;ab1et0;aDy;th; dAEshift;!t41;abCho;aYeTiNoHuG;cky,mp0sh;aKck0f3InJoIuHvGw04y7;ePi5;d,sy,t2V;k0se;e3g;d0nn0;ft0g8HkeKmit3UnJstItGveK;eraGt1;l,te;ab1e9;e0k0;ab1ly;a6Zc2FftJgHtGwd;ab1h7;al,eGiti03;ndaAA;!-1C;bIck0me,nguAOr9Ast,tHuG;gh0n9Y;e,tB;or0;ap53eeKiInoG;ck0wG;ab1i5;ck0ll0ndGss0;!ly;n,p0;aIoHuG;dDi2Ump0stifi2veni1;i9l3;gg2m0;cy,de7gnor0Cll0Bm04nHpso 5FrGss9E;a6releva4; 00aZbr2cYdVfluenUhTi3DnSsOtIvG;eGolE;nt0st0;aAYeIiHoxG;ic9L;ma6;llege4nA6rG;a65esGfeAmed9B;ti5;a26ecuCiItG;aGru62;ll0nt;de,st0;a6er,oce4;abA2ibA2;ce0ti7;iHoG;or;caA9reAJ;lu4QorreAIr4O;deq9Zppropr8Y;fGsitu,vitro;ro4;agi60mKpG;ar2PerfeADoHroG;pBve0;li6rtG;a4Ced;eGi2W;d8Pn9J;!eg7;a4e0;aXeSiPoJuG;ge,mHnGrri0;g8Ut0;b1dr9Oid;lJme3nIok2pe0rrHstGt,ur3;ab1i1;if8X;e93or0;d faFlG;ow;ghHl3re0spG;an8S;!falut62;aHf1ClGsi1M;l0Qp63;dHt0vG;en3y;ab1y;lMnKppJrGte5Z;dHm0roGsh;wi5;! 0Z;e9y;dGgi5;le0so77y;f,low2;aZeXhViUlSoQrIuG;arGe8Yi3Mng ho;ant8Fd1F;aKeJiIoHuG;e70mb7H;ovy,ss,w0;p0s3;at,en,y;b0cHnd,sp0tGy;ef6Mis;ef6L;es0ne,oGve4H;d,fy;ad,ea81ib,oG;b7o02ss0;a4d6Iga6V;aGoG;st3;nGt0;er7Ut1ui0A;inGth8u6Dy,ze0;f6An0;a0Ae04iYlVoNrIuG;ll,nGr7J;!ct16ny;aIeGiend3;eGq3Xsh,tf65;!zi5;cGil,nt7L;tuA;aMcus0lLolKrGul; keeps,ce0eIge2ImGs7Mtuna6;aGer;b1l;caFg1Aign;ish;d0l5D;my;aHoGue4;at0od0w0;g8Jsh0t;erceKguAlJnGrm,t,x0;aHe,iG;sh0te;l,n6F;e0l0;! ;aKd Jeb1mHrG;ti1;a1iniG;ne;up;r0tuA;bl2iOke,lNnMrJst,t,uHvorG;ab1i6;lGx;ty;! G;oGreac5U;ff;cy;l0se;l0nt,r;a0Echo0d73erie,fficie4l0Cm08n01re7OsXthere7vVxG;aTcRerQiFpLtIuG;be7UlG;ta4;e34rG;aGe59;!va5S;a31eHlGo32re0W;ai9;ctHrG;ien64t;ab1ed;ci65t0;eGhanD;l0ss;ct,mi2V;alua6YeGil;n,ryday;sHtG;abli3Qima6V;enG;ti7; Lab58couragKdJgaDhan5Sj0Vri5KsuAtIviG;ab1sG;io9;er0iC;ab1uA;e0i5;rou6sui6;bodi0erDiIpG;hasiGow8ty;ze0;ne4;abo0Qd05eGi6;ct0ga4;gBrLsG;tIyG;!gG;oi5;!eG;rn;ly,n0th3;a0Ve07iQoNrIuG;e,ll,mG;b,p0;aJeHiGop0unk,y;ft0;aGss0;m0ry;g0i9;ab1dDmiHne,ubG;le,t2L;na4;dact55e0ffVp0rTsJvG;eHiGor4W;ne,s1W;ab1r5F;aNcKeas2gruntl2hone58liJmiIordHpl3CtG;a4ingui2Nress2;er3;ss1R;ke0;oHreGu5I;et;uraDv8;gr4XppG;ear0roE;eGty;!ct0;er0icu55; 02aZcVeUfQgenePliNma1AnMpLrang2sItHvoG;ut;ail2ermine06;cribe0erEiHpeMtrG;oi0;gn0re0;a15e14i0Vos4Z;i0se;beHca6ghGv8;tf2X;ra6;eIiG;aGne0;b1nt;at0ct0;m0p;iIlHrG;ea42;aAi0U;de0;dGf,r;!ly,pG;an;facto;i3mInHppBrGs33te0;e0k;ce0gero1Z;agGp;eGi5;ab1d;a16e14h0Wi0Ul0OoPrIuG;dd3r3tG;ab1e;aKeJiIoHuG;ci7el,sh0;ss0wd2;ab1sp;a4Fd44;ck0mp2veGz2;ab1n;ar3Xl09m00nKoJpi0rrIst3uHverGward3zy;ab1t;gh0nt0;e4Ju4N;i5k0l;cTduSfQgrPsMtIvG;eGin36;i0rt06;ai9emIin2PrG;aGibu43ol0;ry,st0;po2O;iHtG;a4itu3Z;d8st0;ue4;iGoGro41;rm0;ct0;eGre6;rn0;biNe0foMmKpG;aAlG;ai9eG;te,x;re0;aGer24on;nd0;rt0;ne0;d,lHoG;ni7ss7;apGectH;seG;ib1;aKeIiHoG;se0u14;ng0p0;aGvB;n,r;im0w0;te0vG;il;aLeIiG;ef,l3;ly;aHck0eGw0;sy;p,t0;llenDnDrG;ge0mi5;a20nt8rtaG;in;gey,lJncIrGt8u1Y;eGri0;ab1f0J;el0;l0mG;!i5;a0Ge05i03lZoQrJuG;dgHrGsy;i0ly,n0;et0;aLiIoHuG;sh0;ad,k0Cwn;ef,gHllGsk;ia4;ht;ke0nd new,sh,ve;aFgNld,mb0na fi0IoFrLtJuG;ncHrgeoG;is;e0y;h8tG;om;ed,i5rG;ow0;us;aHiToGue;ck0nd,oW;ck,nG;d,k;g,ttBzarC;re;aOco0ChiNlLnJst,tGyoN;rHtB;er;ai0;efiGt;ci7t0;ieEoG;ng0v2;nd;r0utifG;ul;ckJd,ld,nIrrHse0wG;dy;en;g0kru1Zn0;!wards; priori,b1Qc1Gd12f0Zg0Uh0Tid0l0Nm0Ln0Fp07r00sTttPutheOvLwG;aJeHkG;wa1X;soG;me;it0re;ant garHeraGoid0;ge;de;nt0I;acIeHractG;ab1i0M;m1Dnd0;h0k0;k0leep,sHtonisG;hi5;eIiHocia13uG;me0;gn0st0;mbGrt0;le0;ab,bitLchKgJrHtifiG;ci7;iEoG;ga4;ue0;ab1iE;raT;pGt;are4eKlJroG;aHprGve0;ia6;ch0;aud0i0;aGtiY;l0r0;gKnounJsw8tiG;cHquG;at2;ipa0G;ce0;ry;aQuG;se0;arKcohIert,iPlHoof,t8;er0;eDow0;olG;ic;mi5;ead;ainJgressiIoniHrG;ee0;zi5;ve;st;foHraG;id;rd0; RdPeNjuFmLo02uKvG;aHerG;se;nc2;ed;lt;it0;st0;pt,qG;ua6;ab1ed0reG;ss0;hoc,infinitG;um;cLhieEknowledDtGu6;ivaHu7;al;te0;ge0;ve0;eIouHuG;ra6se0;nt0;pt0;ando9erQlPoNruMsHunG;da4;e4oHtraGuM;ct;lu6rbi5;ng;te;pt;aGve;rd;aze,e;ra4;nt;nn0;ab1;le","Plural":"ad homineTbNcKdGeEfCgBh9kitMlunchQn8othersQp6rooTs4t0us dollaJweekDyesterd2;h2ics,o0ragediP;d0morrows,nighM;ays;erMingsM;kiLtu0uper bowls,ysteO;d9ffs;a0robleMurposJ;rGt7;othIumbeA;ead starEo0;mFusF;lacie7o9rand slaH;l ozDri0undE;en7;conomiBgDner3v0xamplB;en8;eaths,inne2o0;g0o1;gi7;rs;anari5eil6i1ottag5redit car0;ds;ti3;ank3o1reakfas0;ts;di0ttl0;es;in0;gs;ms","MaleName":"0:A9;1:9J;2:A0;3:9R;4:93;5:7V;6:9X;7:9B;8:8K;9:7H;A:9W;a96b8Kc7Sd6Ye6Af5Vg5Hh4Yi4Oj3Rk3Jl33m25n1Wo1Rp1Iqu1Hr0Xs0EtYusm0vVwLxavi3yDzB;aBor0;cha53h1E;ass2i,oDuB;sEuB;ma,to;nEsDusB;oBsC;uf;ef;at0g;aIeHiCoB;lfga05odrow;lBn16;bDfr9JlBs1;a8GiB;am2Qe,s;e6Yur;i,nde7Zsl8;de,lBrr6y7;la5t3;an5ern1iB;cBha0nce2Wrg7Sva0;ente,t4J;aPeKhJimIoErCyB;!l3ro7s1;av6OeBoy;nt,v4F;bDdd,mBny;!as,mBoharu;a94ie,y;i9y;!my,othy;eodo0Nia6Aom9;dErB;en5rB;an5eBy;ll,n5;!dy;ic84req,ts3Nyl43;aNcottMeLhIiHoFpenc3tBur1Fylve76zym1;anDeBua6A;f0ph8PrliBve4Iwa69;ng;!islaw,l8;lom1uB;leyma7ta;dn8m1;aCeB;ld1rm0;h02ne,qu0Hun,wn;an,basti0k1Nl3Irg3Hth;!y;lEmDntBq3Zul;iBos;a5Ono;!m7Ju4;ik,vaB;d3KtoY;aQeMicKoEuCyB;an,ou;b6dBf67ssel5X;ol2Fy;an,bFcky,dEel,geDh0landAm0n5Dosevelt,ry,sCyB;!ce;coe,s;l32r;e43g3n8o8Hri5C;b7Ie89;ar4Xc4Wha6YkB;!ey,y;gCub6x,yBza;ansh,nal4U;g7DiB;na79s;chDfa4l22mCndBpha4ul,y58;al5Iol21;i7Zon;id;ent2int1;aIeEhilDierCol,reB;st1;re;!ip,lip;d7SrDtB;ar,eB;!r;cy,ry;bLt3Iul;liv3m7LrDsCtBum79w6;is,to;ama,c77;i,l3NvB;il4H;athanIeHiDoB;aBel,l0ma0r2H;h,m;cDiCkB;h5Oola;lo;hol9k,ol9;al,d,il,ls1;!i4;aUeSiKoFuByr1;hamDrCstaB;fa,pha;ad,ray;ed,mF;dibo,e,hamDntCrr4EsBussa;es,he;e,y;ad,ed,mB;ad,ed;cFgu4kDlCnBtche5C;a5Yik;an,os,t1;e,olB;aj;ah,hBk8;a4eB;al,l;hBlv2r3P;di,met;ck,hLlKmMnu4rGs1tCuri5xB;!imilianA;eo,hCi9tB;!eo,hew,ia;eBis;us,w;cDio,kAlCsha4WtBv2;i22y;in,on;!el,oIus;colm,ik;amBdi,moud;adB;ou;aMeJiIl2AoEuBy39;c9is,kBth3;aBe;!s;g0nn5HrenDuBwe4K;!iB;e,s;!zo;am,on4;evi,i,la3Yn5KoBroy,st3vi,w3C;!nB;!a4X;mCn5r10uBwB;ren5;ar,oB;nt;aGeChaled,irBrist40u36y2T;k,ollos;i0Wlv2nBrmit,v2;!dCnBt;e0Uy;a43ri3T;na50rBthem;im,l;aZeRiPoDuB;an,liBni0Ost2;an,o,us;aqu2eKhnJnGrEsB;eChB;!ua;!ph;dBge;an,i;!aB;s,thB;an,on;!ath0n4A;!l,sBy;ph;an,e,mB;!m46;ffGrDsB;sBus;!e;a4AemCmai7oBry;me,ni0H;i5Iy;!e01rB;ey,y;cGd6kFmErDsCvi3yB;!d6;on,p3;ed,r1F;al,es;e,ob,ub;kBob;!s1;an,brahJchika,gHk3lija,nuGrEsDtBv0;ai,sB;uki;aac,ha0ma4;a,vinB;!g;k,nngu3W;nacBor;io;im;aKeFina3RoDuByd42;be1QgBmber3FsD;h,o;m3ra5sBwa34;se2;aEctDitDnCrB;be1Lm0;ry;or;th;bIlHmza,ns,o,rCsBya36;an,s0;lEo3BrDuBv8;hi33ki,tB;a,o;is1y;an,ey;!im;ib;aLeIilbe3YlenHord1rDuB;illerBstavo;mo;aDegBov3;!g,orB;io,y;dy,h43nt;!n;ne,oCraB;ld,rdA;ffr8rge;bri4rB;la1HrBy;eYy;aOeLiJlIorr0CrB;anDedB;!d2GeBri1K;ri1J;cCkB;!ie,l2;esco,isB;!co,zek;oyd;d4lB;ip;liCng,rnB;anX;pe,x;bi0di;arWdRfra2it0lNmGnFrCsteb0th0uge7vBym6;an,ereH;gi,iCnBv2w2;estAie;c02k;rique,zo;aGiDmB;aFeB;tt;lCrB;!h0;!io;nu4;be02d1iDliCm3t1v2woB;od;ot1Bs;!as,j35;!d1Xg29mEuCwB;a1Din;arB;do;o0Fu0F;l,nB;est;aSeKieJoDrag0uCwByl0;ay7ight;a7st2;minEnDugCyB;le;!l9;!a1Hn1K;go,icB;!k;go;an,j0lbeHmetriYnFrEsDvCwBxt3;ay7ey;en,in;moZ;ek,ri05;is,nB;is;rt;lKmJnIrDvB;e,iB;!d;iEne08rBw2yl;eBin,yl;lBn;!l;n,us;!e,i4ny;i1Gon;e,l9;as;aXeVhOlFoCraig,urtB;!is;dy,l2nrad,rB;ey,neliBy;us;aEevelaDiByG;fBnt;fo06t1;nd;rDuCyB;!t1;de;en5k;ce;aFeErisCuB;ck;!tB;i0oph3;st3;d,rlBse;es,ie;cBdric,s0N;il;lEmer1rB;ey,lCroBt3;ll;!os,t1;eb,v2;arVePilOlaNobMrCuByr1;ddy,rt1;aGeDi0uCyB;anDce,on;ce,no;nCtB;!t;d0t;dBnd1;!foCl8y;ey;rd;!by;i7ke;al,lF;nDrBshoi;at,naBt;rdA;!iCjam2nB;ie,y;to;ry,t;ar0Qb0Id0Fgu0Dhme0Cid6jani,lVmTnLputsiKrCsaBu0Dya0ziz;hi;aHchGi4jun,maEnCon,tBy0;hur,u05;av,oB;ld;an,ndA;el;ie;ta;aq;dGgelAtB;hoEoB;i7nB;!iA;ne;ny;reBy;!s,w;ir,mBos;ar;!an,beOeIfFi,lEonDt1vB;aMin;on;so,zo;an,en;onCrB;edA;so;jEksandDssExB;!and3is;er;ar,er;andB;ro;rtA;!o;en;d,t;st2;in;amCoBri0vik;lfo;!a;dDel,rahCuB;!bakr,lfazl;am;allEel,oulaye,ulB;lCrahm0;an;ah,o;ah;av,on","Expression":"aZbYdTeRfuck,gQhKlHmGnFoCpAsh9u7voi01w3y0;a1eKu0;ck,p;!a,hoo,y;h1ow,t0;af,f;e0oa;e,w;gh,h0;! huh,-Oh,m;eesh,hh,it;ff,hew,l0sst;ease,z;h1o0w,y;h,o,ps;!h;ah,ope;eh,mm;m1ol0;!s;ao,fao;a3e1i,mm,urr0;ah;e,ll0y;!o;ha0i;!ha;ah,ee,oodbye,rr;e0h,t cetera,ww;k,p;'3a0uh;m0ng;mit,n0;!it;oh;ah,oo,ye; 1h0rgh;!em;la","Unit":"0:1C;a17b15c0Rd0Qe0Of0Kg0Gh0Din0Cjoule0k06l04mSnRoQpLqKrJsDt8volt0w7y5z4°2µ1;g,s;c,f,n,r1;a,e;b,e2;a0Pb,d0Vo1;tta0J;att0b,ebers;able4b3e2on1sp;!ne0;a2ra0F;!l,sp;spo07;eQiemens0q1; ft,uare 1;c0Jd0If3i0Gkilo0Km1ya0F;e0Nil1;e0li0I;eet0o0E;ankine0eaumur0;t,uart0;a4e2i1ou0Pt;c0Mnt0;rcent,t1;a01roleum barrels;!scal0;hm0uWz;an0HewtS;/s,b,e7g,i3l,m2p1s,²,³;h,s;!²;!/h,cro3l1n;e1li06;! EsD²;g06s0B;gaQter1;! 2s1;! 1;per se1;cond;b,iZm,u1x;men0x0;b,elvins,g,ilo2m1nQ;!/h,ph,²;byYgWmeter1;! 2s1;! 1;per hour;²,³;a,e1g,z;ct1nrys,rtz0;aWogP;al2b,iga9ra1;in0m0;!l1;on0;a3emtOl1oCtG; oz,uid ou1;nce0;hrenheit0rad0;b,xa1;byH;eciCg,l,mA;arat0eAg,l,m9oulomb0u1;bic 1p0;c5d4fo3i2meAya1;rd0;nch0;ot0;eci2;enti1;me4;!²,³;lsius0nti1;g2li1me1;ter0;ram0;bl,y1;te0;c4tt1;os1;eco1;nd0;re0;!s","Place":"aXbWcSdRevergladTfPgNhMiKjfk,kIlGmEneBoAp7que,rd,s6te5u4v3w0yyz;is1y0;!o;!c;a,t;safa,t;nn,x;ask,fo,t,yd;acifSe1h0;l,x;k,nnG;kla,nt,rd;b1w 0;england,hampshire,jersey,mexico;!r;co,i0t,uc;a,nn;a0gw,hr;s,x;an0ul;!s;a0cn,ndianGst;!x;kg,nd,wy;a,reat 0;britain,lak4;co,l0ra;!a;ak,en,fw,ist,xb;al2dg,gk,lt,o1r0t;es;lo,nn;!if;cn,kk,lvd;ms,ntarct2r1tl0ve;!ant1;ct0iz;ic0; ocean","FemaleName":"0:81;1:7E;2:7G;3:7Y;4:65;5:7P;6:7T;7:7O;8:7U;9:7C;A:6K;B:7R;C:6X;D:79;a78b6Pc5Ud5Ce4Rf4Jg47h3Zi3Uj38k2Sl21m1An14o12p0Ur0FsYtNursu9vIwGyEza7;olan2vE;etDon5Z;an2enMhi6PilE;a,la,ma;aHeFiE;ctor1o9rgin1vi3B;l4VrE;a,na,oniA;len5Ones7N;aLeJheIi3onHrE;acCiFuE;dy;c1na,s8;i4Vya;l4Nres0;o3GrE;e1Oi,ri;bit8mEn29ra,s8;a7iEmy;!ka;aTel4HhLiKoItHuFyE;b7Tlv1;e,sEzV;an17i;acCel1H;f1nEph1;d7ia,ja,ya;lv1mon0;aHeEi24;e3i9lFrE;i,yl;ia,ly;nFrEu3w3;i,on;a,ia,nEon;a,on;b24i2l5Ymant8nd7raB;aPeLhon2i5oFuE;by,th;bIch4Pn2sFxE;an4W;aFeE;ma2Ut5;!lind;er5yn;bFnE;a,ee;a,eE;cAkaB;chEmo3qu3I;a3HelEi2;!e,le;aHeGhylFriE;scil0Oyamva2;is,lis;arl,t7;ige,mGrvati,tricFulE;a,etDin0;a,e,ia;!e9;f4BlE;ga,iv1;aIelHiForE;a,ma;cEkki,na;ho2No2N;!l;di6Hi36o0Qtas8;aOeKiHonFrignayani,uri2ZyrE;a,na,t2J;a,iE;ca,q3G;ch3SlFrE;an2iam;dred,iA;ag1DgGliFrE;ced63edi36;n2s5Q;an,han;bSdel4e,gdale3li59nRrHtil2uGvFx4yE;a,ra;is;de,re6;cMgKiGl3Fs8tFyanE;!n;a,ha,i3;aFb2Hja,l2Ena,sEtza;a,ol,sa;!nE;!a,e,n0;arEo,r4AueriD;et4Ai5;elLia;dakran5on,ue9;el,le;aXeSiOoKuGyE;d1nE;!a,da,e4Vn1D;ciGelFiEpe;sa;a,la;a,l3Un2;is,la,rEui2Q;aFeEna,ra4;n0t5;!in0;lGndEsa;a,sE;ay,ey,i,y;a,i0Fli0F;aHiGla,nFoEslCt1M;la,na;a,o7;gh,la;!h,n07;don2Hna,ra,tHurFvern0xE;mi;a,eE;l,n;as8is8oE;nEya;ya;aMeJhadija,iGrE;istEy2G;a,en,in0M;mErst6;!beE;rlC;is8lFnd7rE;i,ri;ey,i,lCy;nyakumari,rItFvi5yE;!la;aFe,hEi3Cri3y;ar4er4le6r12;ri3;a,en,iEla;!ma,n;aTeNilKoGuE;anEdi1Fl1st4;a,i5;!anGcel0VdFhan1Rl3Eni,seEva3y37;fi3ph4;i32y;!a,e,n02;!iFlE;!iE;an;anHle3nFri,sE;iAsiA;a,if3LnE;a,if3K;a,e3Cin0nE;a,e3Bin0;cHde,nEsm4vie7;a,eFiE;ce,n0s;!l2At2G;l0EquelE;in0yn;da,mog2Vngrid,rHsEva;abelFiE;do7;!a,e,l0;en0ma;aIeGilE;aEda,laE;ry;ath33i26lenEnriet5;!a,e;nFrE;i21ri21;aBnaB;aMeKiJlHrFwenE;!dolY;acEetch6;e,ie9;adys,enEor1;a,da,na;na,seH;nevieve,orgi0OrE;ald4trude;brielFil,le,yE;le;a,e,le;aKeIlorHrE;ancEe2ie2;es,iE;n0sA;a,en1V;lErn;ic1;tiPy1P;dWile6k5lPmOrMstJtHuGvE;a,elE;yn;gen1la,ni1O;hEta;el;eEh28;lEr;a,e,l0;iEma,nest4;ca,ka,n;ma;a4eIiFl6ma,oiVsa,vE;a,i7;sEzaF;aEe;!beH;anor,nE;!a;iEna;th;aReKiJoE;lHminiqGnPrE;a,e6is,othE;ea,y;ue;ly,or24;anWna;anJbIe,lGnEsir1Z;a,iE;se;a,ia,la,orE;es,is;oraBra;a,na;m1nFphn0rlE;a,en0;a,iE;el08;aYeVhSlOoHrEynth1;isFyE;stal;ti3;lJnsHrEur07;a,inFnE;el1;a,e,n0;tanEuelo;ce,za;e6le6;aEeo;ire,rFudE;etDia;a,i0A;arl0GeFloe,ristE;a,in0;ls0Qryl;cFlE;esDi1D;el1il0Y;itlin,milMndLrIsHtE;ali3hE;er4le6y;in0;a0Usa0U;a,la,meFolE;!e,in0yn;la,n;aViV;e,le;arbVeMiKlKoni5rE;anIen2iEooke;dgFtE;tnC;etE;!te;di;anA;ca;atriLcky,lin2rItFulaBverE;ly;h,tE;e,yE;!e;nEt8;adOiE;ce;ce,z;a7ra;biga0Kd0Egn0Di08lZmVnIrGshlCudrEva;a,ey,i,y;ey,i,y;lEpi5;en0;a,dNeLgelJiIja,nGtoE;inEn1;etD;!a,eIiE;ka;ka,ta;a,iE;a,ca,n0;!tD;te;je9rE;ea;la;an2bFel1i3y;ia;er;da;ber5exaJiGma,ta,yE;a,sE;a,sa;cFsE;a,ha,on;e,ia;nd7;ra;ta;c8da,le6mFshaB;!h;ee;en;ha;es;a,elGriE;a3en0;na;e,iE;a,n0;a,e;il","FirstName":"aIblair,cGdevFguadalupe,jAk8lashawn,m5r2sh0trinity;ay,e0iloh;a,lby;e1o0;bin,sario;g1ne;ar1el,org0;an;ion,lo;asAe0;ls9nyatta,rry;a1e0;an,ss2;ime,m0n;ie,m0;ie;an,on;as0heyenne;ey,sidy;ndra,ubr0;ey","Month":"aAdec8feb6j2mar,nov8oct1sep0;!t7;!o7;an4u0;l1n0;!e;!y;!r0;uary;!em0;ber;pr1ug0;!ust;!il","Preposition":"'o,-,aKbHcGdFexcept,from,inEmidPnotwithstandiQoCpRqua,sBt6u3vi2w0;/o,hereMith0;!in,oQ;a,s-a-vis;n1p0;!on;like,til;h1ill,o0;!wards;an,r0;ough0u;!oH;ans,ince,o that;',f0n1ut;!f;!to;espite,own,u3;hez,irca;ar1e0y;low,sides,tween;ri6;',bo7cross,ft6lo5m3propos,round,s1t0;!op;! long 0;as;id0ong0;!st;ng;er;ut","Honorific":"aTbScJdIeHgFhon,jr,king,lDmAop,p6queen,r3s0;e1fc,gt,ir,r,u0;ltKpt,rg;cretary,n8;abbi,e0;p6s,v0;!erend;astJhd,r0vt;eside6ince1of0;!essH;!ss;a1essrs,istDlle,me,r0str;!s;j,yD;ady,ieutena0lb,ord,t;nt;en0ov;!eral;sq,xcellency;et,oct7r,utchess;apt7hance5mdr,o0pl;m2ngress0unci4;m0wom0;an;dr,mand0;er;ll0;or;!ain;ldg,rig;d0sst,tty,yatullah;j,m,v","Superlative":"0:18;1:19;a18b12c0Rd0Pe0Mf0Gg0Bh01ic0j00lVmSnQoPpLquiKrIs7t5u3vOw2yel04zan0;et14ind0or1r0;gl0n2;lik00tid0;al0Qhirst0i2oug0F;d0gh10mYn0;aBcar0ex0h8i0Tkinn0l6m5p4t2weE;ee0Ura2;i0Wn0Z;ic0righ0O;al0Ke0P;i2oU;e1gh0Tm0;a3i2;e1n0;d0lPr0M;in0Hlt0vv0;ead0itz0o2uT;om0s0ug01;et0M;hon0l4r2un0;ic2ouP;e0kl0;aine1;il0ran0J;arroGeVice1o2;is0s0;e2o0F;a2lC;sl0t0;a0Bi5o2uck0;ft0n8u3v8w2;e1l0;de1s0;gh07k5v5;o00uic0;a9e6iZo2u08;l3m2t04;el0;lo2;we1;a2ft0;d0v2;enl0ie1;pp0r2;de1;a5h4idd0loom0oof0r2;a0e2isl0oov0;atTie1;asKosK;ie1ud0;a6itQl5oam0rienHu2;lEnn0r2;r0t2;he1;atM;nc0tLult0;a2mpt0;r2s0;l0thl0;amDe2irt0r0ul6;a7eC;age0h8loud0o5r3u2;d5rl0;is9ue2;le1;s3war2z0;dl0;tl0;e3i2;ll0;a2es0;pe1;awd0ig6lo5ounc0ri3u2;rl0s0;gh2;te1;od0;ge1;ngr0;ie1;st","Comparative":"0:16;1:0Z;a15b0Yc0Sd0Oe0Lf0Fg08h02i13j01lWmTnRoQpMquie0JrIs7t5u3vPw2yell04za0H;e0Iin13orse,ri0;g1n2;like1ti11;a0Lhirs0Mi2ouG;di0g0Wme1ni0;aBca10exi0h8il1kin0Bl6m5p4t2wee0D;eep0ra2;ig0Tng0;i0Uright1;a0Fel1;i2ow0;er,g0Pmi0;a3i2;er,ni0;di0llP;dd0int1l0AvU;e4itzi0o2ud0;o00si0u2;gh0;a0Kdd0;hoWl4r2uW;ic2oud0;ei0k1;ain0;i1ld0rang0;arrEeaTo2;i06si0;e2oo0C;a2llB;s1ti0;i5o2ucki0;fUne1u3ve1w2;er,li0;d0si0;g02ke1ve1;ol1ui03;appi0e4il1o2u04;ll2me1tt0;ow0;a2fM;di0v2;en1i0;a7h5idXlooBoofi0r2;ay0e3is1oo2;vi0;a9en0y0;a2o2;st1;i0uR;a6i5la5oa4riend1u2;ll0n2rR;ni0;mi0;tt0;nKtt0ul7;a2mp6;r2si0;li0th1;e4ir3ri0u2;ll0;ti0;ad1ep0;agei0h4louDo3ruel0u2;dd1r1;ld0st1ward1zi0;e2il1;ap0e2;si0;aw7igg0loo7o5rig4u2;r1si0;li0;ht0;ld0un2;ci0;di0;ng2;ri0;er","Gerund":"0:6Z;1:6X;2:6J;3:6Y;4:6W;5:65;6:6B;7:5P;8:58;9:6T;A:6R;B:6U;C:62;D:5S;E:6O;F:6H;G:6L;a6Cb5Wc51d4Ae3Uf3Ig3Bh35i2Uj2Sk2Ql2Hm26n23o1Zp1Hquest44r0Vs05tWuSvPwIyH;aw3ielA;aKe1Zhist5iJoHrec7;n25rH;ki0rD;n3pi0s6t5X;l7nIrHs6tc6vi0;ni0ran1;de2ti0;aIiHo1;eGsE;l6CrD;nHplif1r9si0;dHi1l5R;erHoi0;g3Umi3;aOeNhLoKrIuHw1H;ni0r3;aHeat1ic7us1yi0;di0i3ns46p8ve5;o7r2Zuc6;an7rH;eate3iB;nAs1;ki0l7rg5Bs1;a05c02eZhXiAkip8lWmVnTor1pRtKuHwi04;bmEc7f3Zgg0WppIrHsp3Y;gi0rounAviB;lDor1;aMeLoKrIuH;dDmb5;ai3eHiB;ngthe3s4tc6;p8ri0;mCp8;i3ki0r0Tyi0;it1las6oi5rinH;gi0k5;eHoG;a7e35;as6el5i5o7;amCee8ip8;aHout1;ri0vi0;arc6eIl3GnHrBtt2J;di0si0;ki0mi0;a5Io2rH;aHeG;tc6;i5vi0;a00eJis7oIuH;b4Ni3li0s6;aCl5;aWcVduFfUgTj36lRmQn1pOquNsKtJvHwarA;eHieG;a5li0r4;ai3i2ur3;eIis1or1tH;i0r03;mb5n1rB;es1i2;ai2lHor1resZ;aFyi0;ai3embe2inAoB;aHen1iByi0;ti0xi0;arAr44;l2Su4;al5eiBko3ogni2B;c6li2Aso3;i4nH;gi0ki0;aXerViUlSoQrJuH;bl2Hl5m8rHs6t1;cha4s4K;ai4eKin1oH;cIduFgr51hibEmi4tHvi2C;es1;es4laiC;dIpa2s4vH;en1;ic1;in1p8rt3WsHu2;i0ti0;a3LuH;g9n9si0;c7li0;forCmEsH;is1;c7in1rt4Ds4u4yi0;bJccu2f26pera1utweIverHwi0;coClap8ta7u4whelC;ig6;serBt2X;eIotiH;ci0ng;eAgl1Z;aNeMiKoJuH;ltiplDrH;de2;nito2vi0;nHrro2;di0g5u4;asu2l1nt18r9ssa9;int2Mke d3Qna9rItH;c6te2;kHrD;et1i0;aNeMiJoHum8;aHc7o7;di0ni0;f1ki0mEnIstHvi0;e3i0;i0ki0;ar3c0Ct1;bo2c7uH;g6nc6;ee8iHn2U;c7l5s4;amCoi3uH;d9m8;gno2mPnHss39;cNdica1flu0JhMsKterJvH;eHolB;n1s1;ac1fe2;is1tH;al5r1V;abEibE;luAr1C;agi3pH;or1roB;aLeJi2oIuH;n1rrD;ld fVno2pi0s1;aHl8;di0t1;nd5ppe3rCti0;aMet1los4oLrJuH;arHes4iA;ant35di0;aHip8oGumb5;b24s8;e4v1G;i3t21zi0;aReaQiOlMoIracHunct02;tu2;cu4lJrH;ci0ecHmi0s2X;as1;di0loG;as6oH;at1oAwi0;gu2lHn09xi0;i0li0;ri0tu2;i5l5vo2;ar3chVdEl0AmSnOstNvalua1xH;a0CcLerKis1pHtenA;anAeHl13or1r2Q;c1riH;enF;ci4ti0;e5han9;ablYima1;ab5dJga9hanFjoDric6su2te2vH;isHyi0;io3;i0u2;bodDer9pH;hasiHo25;zi0;oi0;a06eWiLoKrHum8yi0;aIeHif1op8;aCs4;gi0i3;d9i0ub1;fQp8rPsJvH;iHorF;di0ng;aKcIli7mis4plaDtinguH;is6;oHus4;ura9ve2;gr1WppH;ea2roB;ec1;fe2;cNeCfLlive2m0AnDpKsIterH;mi3;cri0RerBiHtroD;g3ri0;ar1enAic1osE;eHi3yi0;at1c1;iAlIrH;ea4;a2i3;nFri0ti0;a09e08h05i1l01oKrHut1;aIeHos4us6yi0;a1dE;c7vi0;llWmSnIo7pDuHve2;g6n1;cPdOfNsLtJvH;eHinF;r1yi0;ai3in0TrH;as1ibu1o5;iHtitu1;de2s1;irCorCron1;uc1;er3;bi3for1i0mJpH;a2lH;ai3;anA;ap4ec1;aJiHo4;n9p8;pi0;iCwi0;aIeH;at1c7;llen9n9r9;a4nte2;l5nce5rHte2u4;i0rD;aUeQlOoLrKuH;dgIrH;ni0yi0;et1;a7us6;as1mIos1rroGtHunF;he2;bi0;oc7;ki0;a2coClJnefEtH;raD;yi0;ieBon9;nHsi0;gi0ni0;bando3c0Cd07fforAgr06iAl02mu4nYppTrQsKttHvoiAwaE;acIeHrac1;mp1nA;hi0ki0;ki0sH;eJiIocia1uC;mi0;g3s1;mb5r1;li0;chiIgHriB;ui0;ng,vi0;eaKlJroH;ac6vi0;hi0;auAyi0;li0ri0;nounFsItH;icipa1;we2;ci0;lIte2;ri0;e9oG;wi0;ei0;dIjus1mEop1;it1;edAi0rH;es4;di0;cJhieBknowled9tiva1;gi0;vi0;ep1oun1u4;si0;ti0;ni0;ng","Determiner":"aAboth,d8e5few,l3mu7neiCown,plenty,some,th2various,wh0;at0ich0;evB;at,e3is,ose;a,e0;!ast,s;a1i6l0nough,very;!se;ch;e0u;!s;!n0;!o0y;th0;er","PastTense":"0:7E;1:6Y;2:74;3:6C;4:73;5:5H;a6Fb5Yc51d45e3Qf3Dg37h30i2Pj2Mk2Il29m1Zn1Wo1Sp1Bques3Er0OsWtNuJvGw7y6;awn0ield0;aBeAhistl0i9o6re65;n7r6;k0ri0;!d1;p0sh0th1;i1Nnt;lk0n7r6sh0t6Mv0;n0ra77;d1t0;a7i6ot0;ew0s4;lu0ri0;n6pli3Trg0s0;d6it0lo5S;er6o0;go0m5;aDe4AhBoAr7u6wi2;n0rn0;a7eat,i6u2;ck0ed;d0in0nsf1p,v5B;ok0rt3Eu67;ank0r6;eat2Hiv0;lk0rg5Bst0;aWcUeQhNkip,lMmLneKoJpHtAu6wit63;bm4ck0ff1gge2pp7r6spe3;g0pr31rou4Rv5X;li0o6;rt0s0;aBeAo9r7u6;di0m5R;a5e6iv0;ngth27ss0t5V;ok,pp0r0;mm0p;in0r0Jy0;it,la50oil0rin6;g0kl0;de,rt0;ak0ez0w;a4We1Pil0ok0;a1Ge1Dip;a7o6;ck0ut;r0v0;ar5Ie8le3n7rv0t6;!tl0;s0t;k0m0;a5Hor0r6;at5Dew0;il0v0;aQe8isk0o7u6;bb0in0l0sh0;am0ll0;aMcLdVfKgJje3lGmFnt0pDquCs9t8v6wa5K;e6iew0;al0l0rs0;a5ir0u3S;e7i2o3Vt6;ed,ri3;m4Wnt0rv0;e2ir0;air0l6o3Rrese5M;ac0i0;a5emb1i3Oov0;a7e5Ji6;ed,v0;t0x0;a56r41;le3us0;a0Ue4Ok5Gogn22;ch0l21s5F;is0n6;g0k0;aLerKiJlGoEr8u6;b1Qll0r6sh0t;ch2Psu0;a1LeAi57o6;c8d7gre51hib4m1Kte2v6;ed,o2N;uc0;e4Ylaim0;di3p27ss0ve52;i51pp0rtr3Us6ur0;ed,t0;a7u6;gg0ng0s0;n0y0;ck0l0;fo2Vm4si2;ck0i4Urt4Dss0us0;b8cc1Eff1per4Qutwei7ver6w0;c3Olap,to3Gus0whelm0;gh0;se20ta5;e7ot6;ed,ic0;ed0gle3;aBeAi8o7u6;ltipli0rd1;nit0Tv0;n6rr0S;d0gl0us0;as12lt0n0Frg0ss1I;inta5ke du0n1Hr7t6;ch0t1;ke6ri0;d,t;aCeBi9o7u6;mp0nk;a6ck0ok0v0;d0n0;ft0k0m4n0ste6v0;d,n0;a27ct0Rt;b0Eck0u6;gh0n3I;e8i7n6;ew,o2T;ck0ll0ss0;pt;a7o5u6;dg0mp0;mm0;gn05mFn6ssu0;cDdic3Sflue3FhCs9ter8v6;e6olv0;nt0st0;a3f1;i2t6;a6ru3;ll0;ab4ib4;l33re14;ag5pr0Q;aAe8ir0o7u6;nt0rri0;ld fa2nSp0st0;a6lp0;d0t;ndl0pp6rm0t0;en0;aAlo39o9r7u6;arante0e38o0V;a6ew,ip,um2K;bb0sp0;t,ve1B;in0th1z0;aGeEiClAo7ractTunc6;ti39;c2Rl7r6;c0eca2m0se0;d0l1U;a1Qew,o6;at,od0;gMl6ni1Ox0;ed,l0;a6ll;r0tJ;il0v6;or0;a0Wcho0d4le3mHnDstBvalu2Sx6;am5c9er7i2p6te06;a0Xerie2Dla5o0Yre2O;c6t0;is0;el0ha1L;ab6im2M;li1B;a1VcourOd8gOha27jVri21s8t7vi6;ed,si2O;!er0;ur0;bodi0erg0p6;has6ow1;iz0;a00eMiBoAr6ump0;a8e7i6opp0;ft0;am0ss0;g0in0;dg0ubt0;d,ed,ff1p,re3s8v6;ed,o6;de,rc0;a9c6lik0mi22pl10tingui0U;o6u21;ur6v1;ag0;g1Spp6;e7r6;ov0;ar0;cFem0fCliv1ma03ni0pAs6;crib0e8i7tr6;oy0;gn0r0;rv0;aZe6i3os4;nt;e7i6;ed,n0;at,ct0;l8o7re6;as0;de;ar0in0;nc0r0t0;aZeYhVit0lSo9r6ut;a7e6i0o1Eu06;at0d4;ck0v0;llNmIn7ok0pi0u6v1;gh0nt0;ceFdu3fDsAt8v6;e6i0X;rt0y0;a5inu0r6;a2ib7ol0;i7tit6;ut0;d1st0;i6o6ro16;rm0;rn0;b5fo8ma7p6;ar0la5;nd0;rt0;in0;aps0e3;a7i6os0;ng0p;im0w0;a7e6;at,ck0w0;lleNng0rg0;as0nt1;ll0me,nc7r6t1us0;ed,ri0;el0;aKeFloEoCr9u6;dg7r6;i0n0;et;o7u6;sh0;ok;a2mb0o2rr6th1u00;ew;ck0;ar0c9l7nef4tr6;ay0;i09o6;ng0;ame;n6s0;g0n0;band08c02dXffoWgVid0lSmRnOppIrFsAtt6void0wa4;ac8e7ra3;ct0;m04nt;h0k0;k0s6;e7i6ociXum0;gn0st0;m6rt0;bl0;ch7gu0r6;iv0;ed,iv0;eaAl8ro6;a6v0;ch0;a6i0;ud0;l0r0;nou7sw1t6;icipJ;nc0;us0;le7t1;er0;g0w;re0;rd0;d8ju2m4oF;it;st0;edd0re6;ss0;c8hi7knowledg0tiv6;at0;ev0;e7ou6us0;nt0;pt0;on0;ed","Adverb":"a06by 04d00eXfShQinPjustOkinda,mMnKoFpDquite,rAs6t3up2very,w1ye0;p,s;ay,ell; to,wards5;h1o0wiN;o,t6ward;en,us;everal,o0uch;!me1on,rt0; of;hVtimes,w05;a1e0;alQ;ndomPthL;ar excellCer0oint blank; Khaps;f3n0;ce0ly;! 0;agYmoS; courFten;ewHo0; longCt withstanding;aybe,eanwhi9ore0;!ovA;! aboR;deed,steS;en0;ce;or1urther0;!moH; 0ev3;examp0good,suF;le;n mas1v0;er;se;e0irect1; 1finite0;ly;ju7trop;far,n0;ow; CbroBd nauseam,gAl5ny2part,side,t 0w3;be5l0mo5wor5;arge,ea4;mo1w0;ay;re;l 1mo0one,ready,so,ways;st;b1t0;hat;ut;ain;ad;lot,posteriori","Actor":"a8br7d4e3gu6k2linkBneedBovert7prov6s1t0;akAe9;e8id9l0t5;eep8;nd7x5;e0iv1;c0pe4;id4;ak3;ddedd2t0;te0;nd0;er","Participle":"a00bXcSdNeHfFgZh06iDkno05lCmAoper06p6r4s1tortLunder0;gone,mU;n1ought,p0urprH;at,rT;o01uck;aEe0;alGcognGl00;a2r0;aBo0;mAvY;id,rtS;anGe0inglW;as9ssF;e3inU;magHn0;dicRflueO;allReat5ig5loPolOra0;ct4;mphas4n2stimOvaluOx0;amCerc0perieK;isN;c3d0g6haIs0;urL;izK;a2e1isc0reamt;our2;cl5f5term5;m0ncGtG;agF;l3o0reD;m0nvi9;b0e;inB;ung;ecome,o0;r0u4;ro5;ctiv5l3n0ssoci5;nou1t0;icip3;nc3;lo0;wn;at0;en","Doer":"aLbIcHdEengineKfCgBhAinstructRjournalNlawyKm9nurse,o8p5r3s1t0;echnEherapM;ailPcientLoldiIu0;pervMrgeon;e0oofG;ceptionIsearE;hotographElumbEoli1r0sychologH;actitionDogrammD;cem8t7;fficBpeH;echanic,inistAus5;airdress9ousekeep9;arden8uard;arm7ire0;fight6m2;eputy,iet0;ici0;an;arpent2lerk;ricklay1ut0;ch0;er;ccoun6d2ge7r0ssis6ttenda7;chitect,t0;ist;minist1v0;is1;rat0;or;ta0;nt","Modal":"c5lets,m4ought3sh1w0;ill,o5;a0o4;ll,nt;! to;ay,ight,ust;an,o0;uld","Conjunction":"aDb9cuz,for,how7in caCno6o5p3supposing,th1vers4wh0yet;eth7ile;eref9o0;!uC;l0rovided that;us;r,therwi6; matt1r;!ev0;er;e0ut;cau1f0;ore;se;lthou1nd,s 0;far as,if;gh","PerfectTense":"have taken","pluPerfectTense":"had taken","FuturePerfect":"will have taken","Demonym":"0:17;a0Wb0Mc0Bd09e08f06g03h01iXjUkSlOmKnHomGpCqatari,rAs6t4u3v2wel0Qz1;am0Eimbabwe0;enezuel0ietnam0G;g8krai11;aiwShai,rinida0Hu1;ni0Prkmen;a3cot0Je2ingapoNlovak,oma0Tpa04udQw1y0X;edi0Jiss;negal0Ar07;mo0uT;o5us0Kw1;and0;a2eru0Ghilipp0Po1;li0Drtugu05;kist2lesti0Qna1raguay0;ma0P;ani;amiYi1orweO;caragu0geri1;an,en;a2ex0Mo1;ngo0Erocc0;cedo0Ila1;gasy,y07;a3eb8i1;b1thua0F;e0Dy0;o,t01;azakh,eny0o1uwaiti;re0;a1orda0A;ma0Bp1;anM;celandic,nd3r1sraeli,ta02vo06;a1iS;ni0qi;i0oneU;aiCin1ondur0unM;di;amCe1hanai0reek,uatemal0;or1rm0;gi0;i1ren6;lipino,n3;cuadoVgyp5ngliIstoWthiopi0urope0;a1ominXut3;niG;a8h5o3roa2ub0ze1;ch;ti0;lom1ngol4;bi0;a5i1;le0n1;ese;liforLm1na2;bo1erooK;di0;a9el7o5r2ul1;gaG;aziBi1;ti1;sh;li1sD;vi0;aru1gi0;si0;ngladeshi,sque;f9l6merAngol0r4si0us1;sie,tr1;a1i0;li0;gent1me4;ine;ba2ge1;ri0;ni0;gh0r1;ic0;an","SportsTeam":"0:1M;1:1T;2:1U;a1Rb1Dc0Zd0Qfc dallas,g0Nhouston 0Mindiana0Ljacksonville jagua0k0Il0Fm02newVoRpKqueens parkJrIsAt5utah jazz,vancouver whitecaps,w3yY;ashington 3est ham0Xh16;natio21redski1wizar12;ampa bay 6e5o3;ronto 3ttenham hotspur;blu1Hrapto0;nnessee tita1xasD;buccanee0ra1G;a7eattle 5heffield0Qporting kansas13t3;. louis 3oke12;c1Srams;mari02s3;eah1IounI;cramento Sn 3;antonio spu0diego 3francisco gi0Bjose earthquak2;char0EpaB;eal salt lake,o04; ran0C;a8h5ittsburgh 4ortland t3;imbe0rail blaze0;pirat2steele0;il3oenix su1;adelphia 3li2;eagl2philNunE;dr2;akland 4klahoma city thunder,r3;i10lando magic;athle0Trai3;de0; 3castle05;england 6orleans 5york 3;city fc,giUje0Lkn02me0Lred bul19y3;anke2;pelica1sain0J;patrio0Irevolut3;ion;aBe9i3ontreal impact;ami 7lwaukee b6nnesota 3;t4u0Rvi3;kings;imberwolv2wi1;re0Cuc0W;dolphi1heat,marli1;mphis grizz3ts;li2;nchester 5r3vN;i3li1;ne0;c00u0H;a4eicesterYos angeles 3;clippe0dodFlaA; galaxy,ke0;ansas city 3nH;chiefs,ro3;ya0M; pace0polis colX;astr0Edynamo,rockeWtexa1;i4olden state warrio0reen bay pac3;ke0;anT;.c.Aallas 7e3i0Cod5;nver 5troit 3;lio1pisto1ti3;ge0;bronc06nuggeO;cowboUmav3;er3;ic06; uX;arCelNh8incinnati 6leveland 5ol3;orado r3umbus crew sc;api5ocki2;brow1cavalie0india1;benga03re3;ds;arlotte horCicago 3;b4cubs,fire,wh3;iteE;ea0ulY;di3olina panthe0;ff3naW; c3;ity;altimore ElAoston 7r3uffalo bilT;av2e5ooklyn 3;ne3;ts;we0;cel4red3; sox;tics;ackburn rove0u3;e ja3;ys;rs;ori3rave1;ol2;rizona Ast8tlanta 3;brav2falco1h4u3;nited;aw9;ns;es;on villa,r3;os;c5di3;amondbac3;ks;ardi3;na3;ls","Country":"0:38;1:2L;a2Wb2Dc21d1Xe1Rf1Lg1Bh19i13j11k0Zl0Um0Gn05om3CpZqat1JrXsKtCu6v4wal3yemTz2;a24imbabwe;es,lis and futu2X;a2enezue31ietnam;nuatu,tican city;.5gTkraiZnited 3ruXs2zbeE;a,sr;arab emirat0Kkingdom,states2;! of am2X;k.,s.2; 27a.;a7haBimor-les0Bo6rinidad4u2;nis0rk2valu;ey,me2Xs and caic1T; and 2-2;toba1J;go,kel0Ynga;iw2Vji2nz2R;ki2T;aCcotl1eBi8lov7o5pa2Bri lanka,u4w2yr0;az2ed9itzerl1;il1;d2Qriname;lomon1Vmal0uth 2;afr2IkLsud2O;ak0en0;erra leoEn2;gapo1Wt maart2;en;negKrb0ychellY;int 2moa,n marino,udi arab0;hele24luc0mart1Z;epublic of ir0Com2Cuss0w2;an25;a3eHhilippinTitcairn1Ko2uerto riM;l1rtugE;ki2Bl3nama,pua new0Tra2;gu6;au,esti2;ne;aAe8i6or2;folk1Gth3w2;ay; k2ern mariana1B;or0M;caragua,ger2ue;!ia;p2ther18w zeal1;al;mib0u2;ru;a6exi5icro09o2yanm04;ldova,n2roc4zamb9;a3gol0t2;enegro,serrat;co;c9dagascZl6r4urit3yot2;te;an0i14;shall0Vtin2;ique;a3div2i,ta;es;wi,ys0;ao,ed00;a5e4i2uxembourg;b2echtenste10thu1E;er0ya;ban0Gsotho;os,tv0;azakh1De2iriba02osovo,uwait,yrgyz1D;eling0Jnya;a2erF;ma15p1B;c6nd5r3s2taly,vory coast;le of m19rael;a2el1;n,q;ia,oI;el1;aiSon2ungary;dur0Mg kong;aAermany,ha0Pibralt9re7u2;a5ern4inea2ya0O;!-biss2;au;sey;deloupe,m,tema0P;e2na0M;ce,nl1;ar;bTmb0;a6i5r2;ance,ench 2;guia0Dpoly2;nes0;ji,nl1;lklandTroeT;ast tim6cu5gypt,l salv5ngl1quatorial3ritr4st2thiop0;on0; guin2;ea;ad2;or;enmark,jibou4ominica3r con2;go;!n B;ti;aAentral african 9h7o4roat0u3yprQzech2; 8ia;ba,racao;c3lo2morPngo-brazzaville,okFsta r03te d'ivoiK;mb0;osD;i2ristmasF;le,na;republic;m2naTpe verde,yman9;bod0ero2;on;aFeChut00o8r4u2;lgar0r2;kina faso,ma,undi;azil,itish 2unei;virgin2; is2;lands;liv0nai4snia and herzegoviGtswaGuvet2; isl1;and;re;l2n7rmuF;ar2gium,ize;us;h3ngladesh,rbad2;os;am3ra2;in;as;fghaFlCmAn5r3ustr2zerbaijH;al0ia;genti2men0uba;na;dorra,g4t2;arct6igua and barbu2;da;o2uil2;la;er2;ica;b2ger0;an0;ia;ni2;st2;an","Pronoun":"'em,elle,h4i3me,ourselves,she5th1us,we,you0;!rself;e0ou;m,y;!l,t;e0im;!'s","Possessive":"anyAh5its,m3noCo1sometBthe0yo1;ir1mselves;ur0;!s;i8y0;!se4;er1i0;mse2s;!s0;!e0;lf;o1t0;hing;ne","City":"a2Pb1Zc1Qd1Le1Kf1Ig1Dh15i13jakar2Ak0Ul0Om0Bn07o05pUquiTrQsJtAu9v6w3y1z0;agreb,uri1T;ang1Ne0okohama;katerin1Crev2X;ars1ellingt1Lin0rocl1;nipeg,terth0S;aw;a1i0;en2Clni2U;lenc2Pncouv0Dr2B;lan bat0Atrecht;a6bilisi,e5he4i3o2rondheim,u0;nSr0;in,ku;kyo,ronFulouB;anj1Yl10miso2Era25; haGssaloni0U;gucigalpa,hr2Jl av0I;i0llinn,mpe26ngi04rtu;chu1Xn2HpQ;a3e2h1kopje,t0ydney;ockholm,uttga0Y;angh1Aenzh1S;o0HvW;int peters0Rlz0Rn 0ppo1A;jo0salvadV;se;eykjavik,i1o0;me,t23;ga,o de janei15;to;a8e6h5i4o2r0ueb1Oyongya1L;a0etor22;gue;rt0zn22; elizabe3o;ls1Erae22;iladelph1Xnom pe07oenix;r0tah tik17;th;lerJr0tr0Y;is;dessa,s0ttawa;a1Flo;a2ew 0is;delTtaip0york;ei;goya,nt0Spl0S;a5e4i3o1u0;mb0Jni0H;nt0scH;evideo,real;n01skolc;dellín,lbour0Q;drid,l5n3r0;ib1se0;ille;or;chest0dalay,i0X;er;mo;a4i1o0vAy00;ndZs angel0D;ege,ma0nz,sbYverpo1;!ss0;ol; pla0Gusan0D;a5hark4i3laipeda,o1rak0uala lump2;ow;be,pavog0sice;ur;ev,ng8;iv;b3mpa0Indy,ohsiu0Fra0un01;c0j;hi;ncheLstanb0̇zmir;ul;a5e3o0; chi mi1ms,u0;stH;nh;lsin0rakliF;ki;ifa,m0noi,va08;bu0QiltC;dan3en2hent,iza,othen1raz,ua0;dalaj0Engzhou;bu0N;eSoa;sk;es,rankfu0;rt;dmont4indhovT;a1ha00oha,u0;blQrb0Dshanbe;e0kar,masc0EugavpiI;gu,je0;on;a6ebu,h2o0raioIuriti00;lo0nstanIpenhagMrk;gEmbo;enn2i1ristchur0;ch;ang m0cago,ttagoK;ai;i0lgary,pe town,rac4;ro;aHeBirminghWogoAr5u0;char3dap3enos air2r0sZ;g0sa;as;es;est;a2isba1usse0;ls;ne;silPtisla0;va;ta;i3lgrade,r0;g1l0n;in;en;ji0rut;ng;ku,n3r0sel;celo1ranquil0;la;na;g1ja lu0;ka;alo0kok;re;aBb9hmedabad,l7m4n2qa1sh0thens,uckland;dod,gabat;ba;k0twerp;ara;m5s0;terd0;am;exandr0maty;ia;idj0u dhabi;an;lbo1rh0;us;rg","Currency":"$,aud,bPcNdIeurHfGgbp,hkd,inr,jpy,kElDp8r7s3usd,x2y1z0¢,£,¥,ден,лв,руб,฿,₡,₨,€,₭,﷼;lotyPł;en,uanO;af,of;h0terlings;e0il4;k0q0;elJ;iel,oubleIp,upeeI;e1ound ster0;lingG;n0soF;ceEn0;ies,y;empi6i6;n,r0wanzaByatB;!onaAw;ori7ranc9t;!o8;en3i2kk,o0;b0ll2;ra5;me4n0rham4;ar3;ad,e0ny;nt1;aht,itcoin0;!s","LastName":"0:31;1:2D;2:2Z;3:28;4:2P;a30b2Sc2Fd27e23f1Yg1Th1Ji1Ej19k12l0Vm0Jn0Fo0Cp02rVsKtFvDwBxAy7zh5;a5ou,u;ng,o;a5eun2Loshi1Eun;ma5ng;da,guc1Smo1Zsh1UzaO;iao,u;a5eb1o4right,u;gn1lk1ng,tanabe;a5ivaldi;ssilj2Xzqu0;a8h7i29o6r5sui,urn1;an,ynisH;lstoy,rr2Lth;at1Nomps2;kah0Rnaka;aCchBeAhimizu,i9mi8o7t6u5zabo;ar0lliv23zuC;al1Nein1;sa,u4;rn3th;lva,mmo1Yngh;mjon3rrano;midt,neid1ulz;ito,n6sa5to;ki;ch0dJz;amAeag1Ti8o6u5;bio,iz,sC;b5dri1HgGj0Pme1XsYux;erts,ins2;c5ve0C;ci,hards2;ir0os;aCe9h7ic5ow1T;as5hl1;so;a5illips;m,n1M;ders1Set6r0t5;e0Kr3;ers;h1Trk1t5vl3;el,te0H;baAg08liveiZr5;t5w1I;ega,iz;a5eils2guy1Kix2owak,ym19;gy,ka5var1E;ji5muU;ma;aDeBiAo7u5;ll1n5ssolini,ñ5;oz;lina,oIr5zart;al1Fe5;au,no;hhail3ll1;rci1s5y1;si;eTmmad3r5tsu04;in5tin0;!o;aAe7i5op0uo;!n5u;coln,dholm;e,fe5n0B;bv5v5;re;rs0Yu;aAennedy,imu9le07o7u6wo5;k,n;mar,znets3;bay5vacs;asV;ra;hn,rl8to,ur,zl3;ansse0Een8ha4imen0o5u4;h5n0Su4;an5ns2;ss2;ki0As0M;glesi8ke7noue,shik6to,vano5;u,v;awa;da;as;aAe8it7o6u5;!a4b1gh0Iynh;a4ffmann,rvat;chcock,l1;mingway,nde5rJ;rs2;ns0ArrOs6y5;asBes;an3hi5;moG;a7il,o6rub1u5;o,tierr0;m0nzal0;nd5o,rcia;hi;er8is7lor04o6uj5;ita;st1urni1;ch1;nand0;d7inste6sposi5vaK;to;in;is2wards;aviBeAi8omin7u5;bo5rand;is;gu0;az,mitr3;ov;lgado,vi;es,s;aEhAlarke,o5;hJl6op1x;er;em6li5;ns;an;an7e6iu,o5ristensEu4we;i,ng,u4w,y;!n,on5u4;!g;mpb6st5;ro;ell;a9e6ha4lanco,oyko,rooks,yrne;ng;ck6ethov5nnett;en;er,ham;ch,h6iley,rn5;es;k,ng;dams,l8nd5;ers5r9;en,on,s2;on;eks6iy7var0;ez;ej5;ev","Duration":"centur4d2hour3m0seconds,week3year3;i0onth2;llisecond1nute1;ay0ecade0;!s;ies,y","Date":"fri4mon4s2t1wed0;!nesd4;hurs2ues2;at0un1;!urd1;!d0;ay0;!s","Holiday":"0:1P;1:1Q;a1Fb1Bc12d0Ye0Of0Kg0Hh0Di09june07kwanzaa,l04m00nYoVpRrPsEt8v6w4xm03y2;om 2ule;hasho16kippur;hit2int0Xomens equalit7; 0Ss0T;aGe2ictor1E;r1Bteran0;-1ax 1h6isha bav,rinityNu2; b3rke2;y 1;ish2she2;vat;a0Ye prophets birth1;a6eptember15h4imchat tor0Vt 3u2;kk4mmer U;a9p8s7valentines day ;avu2mini atzeret;ot;int 2mhain;a5p4s3va2;lentine0;tephen0;atrick0;ndrew0;amadan,ememberanc0Yos2;a park0h hashana;a3entecost,reside0Zur2;im,ple heart 1;lm2ssovE; s04;rthodox 2stara;christma0easter2goOhoJn0C;! m07;ational 2ew years09;freedom 1nurse0;a2emorial 1lHoOuharram;bMr2undy thurs1;ch0Hdi gr2tin luther k0B;as;a2itRughnassadh;bour 1g baom2ilat al-qadr;er; 2teenth;soliU;d aJmbolc,n2sra and miraj;augurGd2;ependen2igenous people0;c0Bt0;a3o2;ly satur1;lloween,nukkUrvey mil2;k 1;o3r2;ito de dolores,oundhoW;odW;a4east of 2;our lady of guadalupe,the immaculate concepti2;on;ther0;aster8id 3lectYmancip2piphany;atX;al-3u2;l-f3;ad3f2;itr;ha;! 2;m8s2;un1;ay of the dead,ecemb3i2;a de muertos,eciseis de septiembre,wali;er sol2;stice;anad8h4inco de mayo,o3yber m2;on1;lumbu0mmonwealth 1rpus christi;anuk4inese n3ristmas2;! N;ew year;ah;a 1ian tha2;nksgiving;astillCeltaine,lack4ox2;in2;g 1; fri1;dvent,ll 9pril fools,rmistic8s6u2;stral4tum2;nal2; equinox;ia 1;cens2h wednes1sumption of mary;ion 1;e 1;hallows 6s2;ai2oul0t0;nt0;s 1;day;eve","Person":"aQbOcKdHemeril lagasse,faGgDhBjk rowling,kAlebron james,m7oprah winfrNparis hiltIr4s3t2uncle,v0womF;a0irgin maB;lentino rossi,n go3;iger woods,yra banks;addam hussaHcarlett johanssEistOlobodan milosevic,omeone,tepA;ay romano,eese witherspoDo1ush limbau0;gh;d stewart,naldinho;an,essiaen,itt romnFo0ubarek;m0thJ;!my;anye west,iefer sutherland,obe bryaG;alle ber0ulk hog3;ry;entlem1irl,rand0uy;fa1mo1;an;thB;ad1enzel washingt0ick wolf,ude;on;!dy;ar1ous0;in;dinal wols0son palm5;ey;arack obama,oy,ro0;!th2;dolf hitl1shton kutch1u0;nt;er","PhrasalVerb":"0:71;1:6P;2:7D;3:73;4:6I;5:7G;6:75;7:6O;8:6B;9:6C;A:5H;B:70;C:6Z;a7Gb62c5Cd59e57f45g3Nh37iron0j33k2Yl2Km2Bn29o27p1Pr1Es09tQuOvacuum 1wGyammerCzD;eroAip EonD;e0k0;by,up;aJeGhFiEorDrit52;d 1k2Q;mp0n49pe0r8s8;eel Bip 7K;aEiD;gh 06rd0;n Br 3C;it 5Jk8lk6rm 0Qsh 73t66v4O;rgeCsD;e 9herA;aRePhNiJoHrFuDype 0N;ckArn D;d2in,o3Fup;ade YiDot0y 32;ckle67p 79;ne66p Ds4C;d2o6Kup;ck FdEe Dgh5Sme0p o0Dre0;aw3ba4d2in,up;e5Jy 1;by,o6U;ink Drow 5U;ba4ov7up;aDe 4Hll4N;m 1r W;ckCke Elk D;ov7u4N;aDba4d2in,o30up;ba4ft7p4Sw3;a0Gc0Fe09h05i02lYmXnWoVpSquare RtJuHwD;earFiD;ngEtch D;aw3ba4o6O; by;ck Dit 1m 1ss0;in,up;aIe0RiHoFrD;aigh1LiD;ke 5Xn2X;p Drm1O;by,in,o6A;r 1tc3H;c2Xmp0nd Dr6Gve6y 1;ba4d2up;d2o66up;ar2Uell0ill4TlErDurC;ingCuc8;a32it 3T;be4Brt0;ap 4Dow B;ash 4Yoke0;eep EiDow 9;c3Mp 1;in,oD;ff,v7;gn Eng2Yt Dz8;d2o5up;in,o5up;aFoDu4E;ot Dut0w 5W;aw3ba4f36o5Q;c2EdeAk4Rve6;e Hll0nd GtD; Dtl42;d2in,o5upD;!on;aw3ba4d2in,o1Xup;o5to;al4Kout0rap4K;il6v8;at0eKiJoGuD;b 4Dle0n Dstl8;aDba4d2in52o3Ft2Zu3D;c1Ww3;ot EuD;g2Jnd6;a1Wf2Qo5;ng 4Np6;aDel6inAnt0;c4Xd D;o2Su0C;aQePiOlMoKrHsyc29uD;ll Ft D;aDba4d2in,o1Gt33up;p38w3;ap37d2in,o5t31up;attleCess EiGoD;p 1;ah1Gon;iDp 52re3Lur44wer 52;nt0;ay3YuD;gAmp 9;ck 52g0leCn 9p3V;el 46ncilA;c3Oir 2Hn0ss FtEy D;ba4o4Q; d2c1X;aw3ba4o11;pDw3J;e3It B;arrow3Serd0oD;d6te3R;aJeHiGoEuD;ddl8ll36;c16p 1uth6ve D;al3Ad2in,o5up;ss0x 1;asur8lt 9ss D;a19up;ke Dn 9r2Zs1Kx0;do,o3Xup;aOeMiHoDuck0;a16c36g 0AoDse0;k Dse34;aft7ba4d2forw2Ain3Vov7uD;nd7p;e GghtFnEsDv1T;ten 4D;e 1k 1; 1e2Y;ar43d2;av1Ht 2YvelD; o3L;p 1sh DtchCugh6y1U;in3Lo5;eEick6nock D;d2o3H;eDyA;l2Hp D;aw3ba4d2fSin,o05to,up;aFoEuD;ic8mpA;ke2St2W;c31zz 1;aPeKiHoEuD;nker2Ts0U;lDneArse2O;d De 1;ba4d2oZup;de Et D;ba4on,up;aw3o5;aDlp0;d Fr Dt 1;fDof;rom;in,oO;cZm 1nDve it;d Dg 27kerF;d2in,o5;aReLive Jloss1VoFrEunD; f0M;in39ow 23; Dof 0U;aEb17it,oDr35t0Ou12;ff,n,v7;bo5ft7hJw3;aw3ba4d2in,oDup,w3;ff,n,ut;a17ek0t D;aEb11d2oDr2Zup;ff,n,ut,v7;cEhDl1Pr2Xt,w3;ead;ross;d aEnD;g 1;bo5;a08e01iRlNoJrFuD;cDel 1;k 1;eEighten DownCy 1;aw3o2L;eDshe1G; 1z8;lFol D;aDwi19;bo5r2I;d 9;aEeDip0;sh0;g 9ke0mDrD;e 2K;gLlJnHrFsEzzD;le0;h 2H;e Dm 1;aw3ba4up;d0isD;h 1;e Dl 11;aw3fI;ht ba4ure0;eInEsD;s 1;cFd D;fDo1X;or;e B;dQl 1;cHll Drm0t0O;apYbFd2in,oEtD;hrough;ff,ut,v7;a4ehi1S;e E;at0dge0nd Dy8;o1Mup;o09rD;ess 9op D;aw3bNin,o15;aShPlean 9oDross But 0T;me FoEuntD; o1M;k 1l6;aJbIforGin,oFtEuD;nd7;ogeth7;ut,v7;th,wD;ard;a4y;pDr19w3;art;eDipA;ck BeD;r 1;lJncel0rGsFtch EveA; in;o16up;h Bt6;ry EvD;e V;aw3o12;l Dm02;aDba4d2o10up;r0Vw3;a0He08l01oSrHuD;bbleFcklTilZlEndlTrn 05tDy 10zz6;t B;k 9; ov7;anMeaKiDush6;ghHng D;aEba4d2forDin,o5up;th;bo5lDr0Lw3;ong;teD;n 1;k D;d2in,o5up;ch0;arKgJil 9n8oGssFttlEunce Dx B;aw3ba4;e 9; ar0B;k Bt 1;e 1;d2up; d2;d 1;aIeed0oDurt0;cFw D;aw3ba4d2o5up;ck;k D;in,oK;ck0nk0st6; oJaGef 1nd D;d2ov7up;er;up;r0t D;d2in,oDup;ff,ut;ff,nD;to;ck Jil0nFrgEsD;h B;ainCe B;g BkC; on;in,o5; o5;aw3d2o5up;ay;cMdIsk Fuction6; oD;ff;arDo5;ouD;nd;d D;d2oDup;ff,n;own;t D;o5up;ut"}
+module.exports={"Noun":"0:2K;1:1V;a27b1Yc14d0Ue0Nf0Mg0Jh0Di05journ1lZmQnews servi2EoOpIquartet,rFs6t4u2vs,worke1F;ni2tilD;on,v7;ce,ele2im2Hribun1v;communica1Egraph,v0S;av0Achool,e8o5t2ubcommit1Fyndicat;ate 2ockP;poli27univ2;ersi23;ci2ns;al club,et2;e,y;cur2rvice0;iti27;adio,e2;gionLs2;ourc24tauraS;artners6d,e5harmac4izza,l3ol2roduc11;i1Wytechn0A;!c;a0Geutical0;ople's par1Qtroleum;!hip;bservHffi1Rptic1r2;chestra,ganiz1Z;a7d,e5i3otor0use2;e,um;lita1Bnist2;e08ry;dia,mori1rcantile2; exchange;!ch1Ngazi3nage06r2;i2ket0D;ne;ab5i4oc2;al 2;aIheaH;beration army,t;or2s;ato0Z;e,n2;dustri1Fs5ter4vest2;me2o09;nt0;nation1sta11;titut2u14;!e2;! of technology,s;e4o2;ld2sp0Itel0;ings;a2ra5;lth a2;uth0T;a3ir0AovernJroup,ui2;ld;s,zet0P;acul0QedRm,ound12und,y;g7lectr6n5s4t3veningHx2;!p;at,c;p,ta0K;semb01terprise0;ic;!li05;a9e4i2rillG;rectora0Fv2;is0T;part2velop5;e4ment2;! sto2s;re;ment;ily2ta; news;aRca,entrQhNircus,lLo2rew;alit0KlHm9n3rp2unc6;!or0Is;fe5s2taine9;e3ult2;ing;il;d2ren04;er0C;m4p2;any,rehensAute2;rs;i4uni2;ca2ty;tions;ss06t2;tee;!le2;ct2ge;ive;ini2ub;c,qI;emica3oir,ronic2urch;le;ls;al bank,e;!fe,is4p2thedr1;it1;al;se;a7c,l,o6r3u2;ilding socieDreau;ands,ewe3other2;hood,s;ry;ard,ys;!n2;k,q2;ue;dGgencDirBrAss6ut2viH;h3ori2;te;ori2;ty;oc4u2;ran2;ce;!i9;c,my;l2ways;in3;e0i2y;es;!s;!ministr2;at2;ion","Uncountable":"0:1B;a1Gb1Ac13e0Xf0Rg0Nh0Hi0Ej0Dknowled1Fl08mZnYoXpSrPsCt8vi7w1;a5ea0Bi4o1;o2rld1;! seJ;d,l;ldlife,ne;rmth,t0;neg7ol09;e3hund0ime,oothpaste,r1una;affSou1;ble,sers,t;a,nnis;aBceVeAh9il8o7p4te3u1;g1nshi0M;ar;am,el;ace2e1;ciOed;!c10;ap,cc0ft0B;k,v0;eep,opp0K;riJ;d07fe0Ul1nd;m0Ot;ain,e1i0U;c1laxa0Csearch;ogni0Brea0B;a4e2hys0Elast9o1ress00;rk,w0;a1pp0trol;ce,nR;p0tiK;il,xygen;ews,oi0A;a7ea5i4o3u1;mps,s1;ic;nHo06;lk,st;sl1t;es;chi1il,themat00;neD;aught0e3i2u1;ck,g05;ghtnVqu0AteratI;a1isH;th0;ewel7usti06;ce,mp1nformaOtself;ati1ortan04;en03;a4isto3o1;ck1mework,n1spitaliZ;ey;ry;ir,lib1ppi9;ut;o2r1um,ymnastJ;a7ound;l1ssip;d,f;i5lour,o2ruit,urnit1;ure;od,rgive1wl;ne1;ss;c6sh;conom9duca5lectriciKn3quip4th9very1;body,o1th7;ne;joy1tertain1;ment;tiA;a6elcius,h4iv3loth2o1urrency;al,ffee,nfusi8tt8;ing;ics;aos,ee1;se;ke,sh;a3eef,is2lood,read,utt0;er;on;g1ss;ga1;ge;c4dvi3irc2mnes1rt;ty;raft;ce;id","PresentTense":"0:7F;1:7G;2:7H;3:73;4:67;5:5X;6:77;7:5P;8:47;a6Ob66c56d4De3Vf3Kg3Dh37i2Vj2Tk2Ql2Hm26n23o1Zp1Equest46r0RsYtOuKvHwAy9;aw2iel6;aCe1Zhistl0iBo9rec4;n25r9;ks,ri0;ns,p0sh0t68;l4nAr9sh0t6Wv0;ns,ran1;de3ts;aAi9ot0;e73si1;lu0ri0;n9plif1rg0s0;d9it0l61;er9os;g3Xm5D;aHeGhEoDrAu9w1H;n0r2;aAea1i9us1;c4es;d0i2ns48ps,ve5;rt3Mu6H;an4r9;eate2iv0;n6s1;k0l4r5Kst0;a00cXeThQid0ki8lPmOnMor1pKtDu9wit6C;bmi1c4f41gg0TppAr9sp40;g0pr38roun6v66;li0o9;r1s0;aEeDoCrAu9;di0m60;ai2e9iv0;ngthe2ss0t64;ps,r0;ms,ps;i2k0r0Qys;i1la59oi5rin9;gs,kl0;e9o69;a4ez0;a55el5il0ok0;a7ee8i8;aAo9;c4u1;r0v0;ar5ReBl3GnArv0t9;s,tl0;ds,s0;ks,ms;aAor0r9;at5Me5X;l0r0;i5v0;aTeBis4oAu9;bs,i2l0sh0;a7l5;aPcOdZfNgMj35lJmIn1pGquFsCtBv9war6;e9ie5Q;a5ls,rs0;ai2ir0ur2;eAis1or1t9;rYs;m54n1rv0;es1ir0;ai3l9or1resU;ac0i0;ai2embe3in6ov0;aAen1i9;es,v0;t0xs;ar6re1;l2Qus0;al5e4Wko2ogn29;ch0l28so2;is0n9;g0ks;aSerQiPlMoKrBu9;b1Wl5m8r9sh0ts;ch2Usu0;a1ReEin1o9;cCdBgre58hi0Tm1QtAv9;es,id0;es1;uc0;e54lai7;dApar0ss0v9;en1;ic1;in1ps,rt3Zs9u3;es,ts;aAu9;gs,ng0;ns,ys;c4l0;for7mi1s9;is1;c4in1rt4Gss0us0ys;bBccu3f21per4TutweAver9w0;c3Tla8tak0us0whel7;ig2J;se22t2W;eAot9;es,ic0;e6gl1U;aFeEiCoBu9;ltipli0r9;de3;nito3v0;n9rro3;ds,gl0us0;as13l1nt12rg0ss21;int2Lke du0n20rAt9;ch0te3;k9ri0;e1s;aFeEiBo9um8;a9c4o4v0;ds,ns;f1k0mi1nAst9v0;e2s;es,ks;ar2ct0Sts;bo3c4u9;g1Vn3L;ee8iAno9;c4ws;c4l5ss0;a7oi2u9;dg0m8;gnor0mIn9ssu0;cGdic3Wflue3KhEsCterBv9;e9olv0;n1s1;ac1fer0;is1t9;al5r1T;a9i9;bi1;lud0re16;ag1Vp9;or1r0S;aDeBir0oAu9;n1rri0;ld fMno3p0s1;a9l8;ds,ts;ndl0ppe2r7t0;aEe1lo3DoDrBu9;ar9e3Cid0;ante0ds;a9i8o35um2O;bs,s8;es,v1D;i2t22z0;aIeaHiFlDo9ractVunctU;c2WlBr9;c0ec9ms,se0;as1;ds,lo2X;a1Uo9;a1o6ws;gOl9ni1Sxs;es,ls;rs,tM;i5l5vo3;ar2chPdi1l04mMnHstFvalu2Xx9;am14cDerBis1p9ten6;an6e9l11or1re2T;c1rie2I;c9ts;is0;e5ha18;ab9im2Q;li1G;a20cour09dCg09ha2Cjo1Mri26sBte3vi9;es,s9;io2;ur0;s,ur0;bodi0erg0p9;has9o25;iz0;os;aZeOiEoDr9um8;aBeAif1o8;ps;a7ss0;gs,i2;dg0es,ub1;es,fHps,rGsAv9;es,id0orc0;aBc9lik0mi25pla15tingui0Y;o9u24;urQve3;g1Ypp9;ea3r9;ov0;ec1;fe3;cGe7fDlive3m06ni0pCs9term07;crib0eAi9tro0W;g2r0;rv0;ar1en6ic1osi1;eAi9;es,n0;a1c1;id0lAre9;as0;ar0in0;m9nc0r0t0;ag0;a05e04h00it0lWoCr9u1;aAe9i0o1Ju0C;at0di1;c4v0;llRmMnBo4pi0u9ve3;g9n1;hs;cIdHfGsDtBv9;e9i12;r1ys;ai2inu0r9;as1ibAo5;iAtit9;ut0;de3s1;ir7or7ron1;uc1;er2;bCes,for1mBp9;ar0l9;ai2;an6;in0;aps0ec1;aAi9os0;nXps;i7ws;ms;aAe9;a1c4ws;lle9ng0rg0;ng0;as0nte3;l5nce5r9te3us0;es,ri0;ls;aOeIlGoErCu9;dAr9;i0ns;ge1;ak0u9;sh0;as1mbs,os1rro09t9u04;he3;oc4;ks;a3cDlBnefi1t9;ra9;ys;i0Aon9;gs;om0;n9s0;gs,s;bando2c03d00ffor6gYi6lUmTnPppKrHsCtt9voi6wai1;acAe9rac1;mp1n6;h0ks;ks,s9;eAi9ociZum0;g2s1;m9r1;bl0;chAgu0r9;iv0;es,iv0;eaClBro9;a9v0;ch0;au6i0;ls,rs;nouBsAt9;icipM;we3;nc0;us0;lAte3;rs;eg0o9;ws;re0;ds;d9jus1mi1op1;re9s;ss0;cBhiAknowledg0tiv9;at0;ev0;ep1oun1us0;es;ts;ns","Infinitive":"0:6B;1:6L;2:6O;3:6G;4:6M;5:5Q;6:4W;7:6K;8:5Y;9:69;a5Xb5Jc4Nd3We3Hf34g2Yh2Si2Hj2Fk2Cl23m1Un1Ro1Np18ques35r0MsWtNuJvHwByA;awn,ield;aCe1Nhist8iBoAre5O;nd0r1Z;n,pe,sh,th0;lk,nBrAsh,t68ve;n,ra6S;d0t;aAi0Jo7;l60ry;nApli3Pr1se;dAi7lo5F;erAo;go,mi4V;aHeGhEoDrBuAwi3;ne,rn;aAi5Au3y;de,in,nsf0v51;r2Ou5V;ank,rA;eat2Di4;nd,st;ke,lk,s7;aYcVeShQi3YlPmOnMo4KpKtEuAwit5Q;ck,ff0gge3ppBrAspe5;ge,pri2rou4Hvi4;ly,oA;rt,se;aEem,oDrBuA;dy,mb8;a4AeAi4;ngth22ss,t5I;p,re;in,ke,r0Ky;la4Ooil,rinA;g,k8;eAow;ak,e2U;a4Ke1Mi8o4L;am,e1B;aAo4M;re,ve;ar57eBle5nAr4tt8;d,se;k,m;aBo6rA;at53ew;le,re;il,ve;aTeCisk,oBuA;b,in,le,sh;am,ll;aPcOdu9fNga57je5lLmKnt,pIquHsEtDvAwa57;eBiA;ew;al,l,r2;a3Ki6u3H;eBi3o3MtA;!ri5;mb8nt,r4;e3i6;air,lAo3Irese59;a9y;a3Demb0i3Fo4;aAe56i4y;te,x;le5u2;a0Tei4k55ogni20;ch,li1Zs54;i2nA;ge,k;aNerMiLlJoHrCuA;b1Pll,mp,rAsh;cha2s48;ai2eDi4WoA;cBdu9gre4Omi2te3vA;e,i2H;e4Mlaim;di5pa6ss,ve4S;i4Rp,rtr3LsAur;e,t;a3BuA;g,n1;ck,le;fo2Osi3;ck,i4Lrt44ss,u2y;bCccur,ff0pera7utweBverAwe;co3Pta38u2whelm;igh;ser4ta2N;eBotA;e,i9;ed,gle5;aFeEiCoBuA;ltiply,rd0;nit0Tve;nArr0S;d,g8us;asu6lt,n0Fr1ssa1;inta2Dke d3Ina1rBtA;ch,t0;k,ry;aGeFiCoAu13;aAck,ok,ve;d,n;ft,ke,nBstAve;!en;e,k;a21c06;b0Fck,uA;gh,n3B;eBiAno2A;ck,ll,ss;ep;am,o1YuA;d1mp;gno6mInAss31;cGdica7flu0CsDterCvA;eAol4;nt,st;a5fe6;i3tA;aAru5;ll;lu15r13;agi1RpA;o1Pro4;aDeCi6oBuA;nt,rry;ld fa3nVpe,st;ad,lp;nd8ppArm,te;en;aEet,lo32ove1CrCuA;arAe31i0V;ant2Vd;aAow,umb8;b,sp;in,th0ze;aKeaJiHlFoCracBuncA;ti34;tu6;cus,lBrA;ce,eca3m,s2N;d,l1P;a1LoA;od,w;gu6lAni1Jx;e,l;r,tu6;il,ll,vA;or;a0Tcho,le5mLnIstGvalua7xA;a03cEerDi3pAte0X;a0WeAla0Uo0Xre2H;ct,riA;en9;ci2t;el,han1;abAima7;li16;ab8couRdBga1han9jXri1Wsu6t0vA;isi2Iy;!u6;body,er1pA;hasiAow0;ze;aZeOiGoFrBuA;mp;aCeBiAop;ft;am,ss;g,in;!d1ubt;e,ff0re5sBvA;e,iQor9;aDcAli0Rmi1Vpl0Wtingui0Q;oAu1U;uAv0;ra1;gr1MppA;ear,ro4;cGem,fFliv0ma04ny,pEsBterA;mi05;cribe,er4iBtrA;oy;gn,re;a00eZi5;e5i00y;iClBrA;ea2;a6iX;de;ma1n9re,te;a02e01hYi7lWoCrA;aAea7o1Cu07y;ck,ve;llSmLnBok,py,uAv0;gh,nt;ceIdu5fGsEtCvA;eAin9;rt,y;aIin0NrA;a3ibu7ol;iAtitu7;d0st;iAoAro18;rm;rn;biFfoEmaDpA;a6laA;in;re;nd;rt;ne;ap2e5;aAiTo2;im,w;aBeA;ck,w;llen1n1r1;a2nt0;ll,ncBrAt0u2;e,ry;el;aLeHloGoErBurA;n,y;aBuA;sh;ke;a3mb,o3rrAth0un9;ow;ck;ar,lBtrA;ay;ie4oA;ng;nAse;!g;band0Cc04d00ffoZgrYid,lVmu2nSppNrKsEttAvoid;acCeBra5;ct;m08nd;h,k;k,sA;eCiBocia7uA;me;gn,st;mb8rt;le;chBgAri4;ue;!i4;eaDlCroA;aAve;ch;aud,y;l,r;noun9sw0tA;icipa7;ce;lBt0;er;e1ow;ee;rd;dBju3oK;st;!reA;ss;cDhie4knowled1tiva7;te;ge;ve;eCouBu2;se;nt;pt;on","Organization":"0:43;1:3K;a38b2Pc29d20e1Xf1Tg1Lh1Gi1Dj19k17l13m0Tn0Go0Dp07qu06rZsStFuBv8w3y2;amaha,m0Youtu2Rw0Y;a4e2orld trade organizati40;lls fargo,st2;fie22inghou17;l2rner br3C;-m12gree30l street journ25m12;an halOeriz3Visa,o2;dafo2Gl2;kswagMvo;bs,n3ps,s2;a tod2Qps;es34i2;lev2Wted natio2Tv; mobi2Jaco beQd bNeBgi fridaAh4im horto2Smz,o2witt2V;shiba,y2;ota,s r Z;e 2in lizzy;b4carpen32daily ma2Wguess w3holli0rolling st1Ms2w3;mashing pumpki2Nuprem0;ho;ea2lack eyed pe3Eyrds;ch bo2tl0;ys;l3s2;co,la m13;efoni09us;a7e5o3pice gir6ta2ubaru;rbucks,to2M;ny,undgard2;en;a1x pisto2;ls;few24insbu25msu1W;.e.m.,adiohead,b7e4oyal 2yan2W;b2dutch she5;ank;/max,aders dige1Ed 2vl31;bu2c1Thot chili peppe1lobst27;ll;c,s;ant2Uizno2E;an6bs,e4fiz23hilip morrCi3r2;emier26octer & gamb1Qudenti14;nk floyd,zza hut;psi27tro2uge09;br2Pchina,n2P; 3ason1Wda2F;ld navy,pec,range juli3xf2;am;us;aBbAe6fl,h5i4o2sa,wa;kia,tre dame,vart2;is;ke,ntendo,ss0K;l,s;stl4tflix,w2; 2sweek;kids on the block,york09;e,é;a,c;nd1Ss3t2;ional aca2Do,we0P;a,cYd0N;aAcdonald9e6i4lb,o2tv,yspace;b1Lnsanto,ody blu0t2;ley crue,oro0N;crosoft,t2;as,subisO;dica3rcedes-benz,talli2;ca;id,re;'s,s;c's milk,tt12z1X;'ore09a4e2g,ittle caesa1td;novo,x2;is,mark; pres6-z-boy;atv,fc,kk,m2od1J;art;eopardy,iffy lu0Ko4pmorgan2sa;! cha2;se;hnson & johns1Rom06y d1Q;bm,hop,n2tv;c,g,te2;l,rpol; & m,asbro,ewlett-packaTi4o2sbc,yundai;me dep2n1I;ot;tac2zbollah;hi;eneral 7hq,l6o3reen d0Hu2;cci,ns n ros0;ldman sachs,o2;dye2g0A;ar;axo smith kliZencore;electr0Hm2;oto1;a4bi,da,edex,i2leetwood mac,oGrito-l09;at,nancial2restoV; tim0;cebook,nnie mae;b05sa,u,xxon2; m2m2;ob0G;aiml0Ae6isney,o4u2;nkin donuts,po0Wran dur2;an;j,w j2;on0;a,f leppa4ll,p3r spiegZstiny's chi2;ld;eche mode,t;rd;aFbc,hCiAnn,o4r2;aigsli6eedence clearwater reviv2;al;!ca co6l5m2o1st05;ca3p2;aq;st;dplMgate;la;a,sco2tigroup;! systems;ev3i2;ck fil-a,na daily;r0Hy;dbury,pital o2rl's jr;ne;aGbc,eCfAl6mw,ni,o2p,rT;ei4mbardiKston 2;glo2pizza;be;ng;ack & deckGo3ue c2;roY;ckbuster video,omingda2;le; g2g2;oodriO;cht4e ge0n & jer3rkshire hathaw2;ay;ryI;el;nana republ4s2xt6y6;f,kin robbi2;ns;ic;bYcTdidSerosmith,ig,lMmGnheuser-busFol,ppleBr8s4t&t,v3y2;er;is,on;hland3s2;n,ociated G; o2;il;by5g3m2;co;os; compu3bee2;'s;te1;rs;ch;c,d,erican4t2;!r2;ak; ex2;pre2;ss; 5catel3t2;air;!-luce2;nt;jazeera,qae2;da;as;/dc,a4er,t2;ivisi2;on;demy of scienc0;es;ba,c","Singular":"ad hominRbNcLdHeFfDgCh9kitMlunch,n8others,p6room,s4t0us dollar,weekEyesterd2;h2ic,o0ragedy;d0morrow,night;ay;ere,iM;ky,tu0uper bowl,ystM;dAff;a0roblKurpo4;rt,t8;othiHumbB;ead start,o0;me,u0;se;laci7od,rand slam;ri0undiC;end;conomy,gg,ner3v0xamp9;ent;eath,inn2o0;g0or;gy;er;anary,eili4i0ottage,redit card;ty;anki2o0reakfast;dy,tt0;le;ng;em","Adjective":"0:B0;1:95;2:BF;3:BL;4:B7;5:BE;6:6S;7:7U;8:A2;9:AY;A:9P;B:9U;C:9S;D:AK;E:AV;aABb9Hc8Nd80e7Ef6Kg64h5Ki4Wj4Uk4Sl4Dm3Yn3No35p2Kqu2Hr24s0Vt0Ku01vWwHyFza2Z;eFou5;ar1ll5N;aReOhMiIoGrF;o5y;man1oFuA0;dA3zy;ck0deHlGnBry,sF;e,tfC;d,y;!sprAU;i2oleF;!sa6;aGek1iBJlFste7Lt;coA1l o6Y;k,ry;rFst8Z;m,p0y;aHeGiFola6Mulgar;bBFce versa,ol7T;ng8VrB4xi5;ca3gFin,nilla,st;abo9Jue;g1ltWnKpHrb7XsGtterF;!moAK;efCu4; f4Nb2ScoADpGsFt2J;ca6et,ide dMtairs;er,i6Y;aObecoAAconvin3EdeMeLfair,ivers4knKlike1precede01rIsGtiBwF;ielBritt9F;ig2YuF;pervis0spec49;eFu1;cognKgul9Zl9Z;own;ndi5v99xpect0;cid0rF;!grou93stood;ppeaOssuA0uthorF;iz0;i3Xra;aMeKhIiHoGrF;anqu81enBi2oubl0ue;gethAp,rpEugh;dy,g8Fme1ny,r0;iFough77;ck,n,rs6F;ena96lFm7Ln9rr4C;li5;boo,lFme,n,rt,st81wdD;eFl;nt0;a0Mc0Je0Hh0Fi0Dkin1Il09m07nobb5Eo02pZquYtMuGwFymb9K;e6Ui2T;bIdd8Qi gener89pGrF;e,re4;erFre8U;! dupAb,fi93i3N;du0sFt6;eq7Btan5Y;aPeMiLrGuFyl56;pErB;aGiFo5;ct,p0;ightFn8P;! fFfF;or8I;ff,ll;adGep,rF;eotyp0n;fa99y;le,nda9Wrk;a8eam4U;a8eciGiFlendEot on,rig1R;cy,r2I;al,f8Y; call0ci4ft,lHmbAot8Hphist2WrGuFvi67;r,th5R;dEe,ry;eFid;!mn;aFel1o20ug;ll,rt;eHiFow,y;ck,g7BmF;!y;ek,ndA;ck,gnifica3l1mp6nF;ce8g6;aFeAi0Ho8Iri0Sut,y;dy,ll36m6Wrp;cFda2em1lf4Bni2Ppa92re7ve8xy;o23r5Ru8;aFient32reec80;rFt7Z;ce,y;cr0d,fe,int1l51me,ns serif,v3I;aPeJiHoFu7T;bu8Hma7To45sy,tt7IuFy4;gh,nd,ti7;ch,ghtFpe,tzy,v4;!-35fC;aJbel,condi2dIfres7RlHmo2public5PsFtard0;is62oF;lu2na3;e20uc60;!un8P;dy,l,r;bEciFgg0nd6TpEre,w;al,st;aGiF;ck,et;ck,i3;aWeUhoTiSlQoPrHuF;bl7Vny,rF;e,p6;eKiHoF;!b0FfFmi4HpAud;ou6Uu9;cGmFor,sti7va2;aDe;ey,k1;ci9g4QmF;a0Pi83;li2or,puli7Rt44;aFeasa3;cEin,nt;cayu7nk;ny;ac5TrFti2;fe83ma44plexi5ver9;i2Vle,rFst;allel0tF;-ti6Pi4;bSdd,ffQi1kPldOnNpMrLthAutKvF;al,erF;!aHnGt,wF;eFroug5U;ig5T;ll;do1Cer,g3Zsi6M;an6Kg1T;en,posi2; boa7Wg3Wli7;! fashion0;!ay; gua7TbFi6Mli7;eat;eHsF;cFer0Sole2;e7u8;d5Fse;aOeLiJoFua74;b6isy,nHrGsy,tF;ab6;m4th3L;chala3descri7Hstop;ce,gFmb6;ht1;aFcessaDighbor1w,xt;rFt;!by;i6Ok0rr16;aPeMiJoGuF;ltip6nda7te;bi6derGist,lBnth1oFst;dy,t;a2n;ld,nF;iaFor;tu8;aGdi6Rek,ll0XnaFre;ci5;n,s1ty;cHd,genta,in,j0EkeshiGle,mmFny,sculi7tu8;oth;ft;ab8ho;aReNiJoGuF;cky,sh;f2PnGo9uFve1wSy4;d,sy,t1T;e1g;g4Kke1mHtFve1;eraFt6;l,te;it0;ftHgFth4wd;al,eFitiT;ndaD;!-0O;me,nguEr56st,tF;e,tA;ap31e4VindF;!ly;agg0ol1uF;i4Cstifi0veni6;cy,de4gno6Gll01mWnHpso 33rF;a2releF;va3; RaQbr0corPdNflu20i21nMsLtF;a64eHiGoxF;ic5A;ma2;llege3n9rF;esFmed55;ti5;a7ecu8i4Rta3;a2er,oce3;iGoF;or;re5U;deq5Kppr4X;fFsitu,vitro;ro3;mHpF;ar1Lerfe5PoFropA;li2r2W;eFi1O;d4Rn9;!eg4;aReNiKoHuF;ge,mFngD;b6dr5Bid;llGme1ne4Zok0rrFs0Vt,ur1;if4S;ow;ghGl1spF;an4P;!falut2Z;aGf1BlFsi2J;l0FpfC;dy,vF;en1y;lKnIppy,rFt2W;dGroFsh;wi5;! 0L;dFgi5;so3Ry;f,low0;aSenRhPiOlMooLrFung ho;aIeHis1oFue3N;oFss;vy;at,en,y;c2Knd,tFy;efCis;d,fy;ad,ea45ib,oF;b4oX;a3dBga3K;aFoF;st1;er3Zt6ui7;iFuBy;nfC;a02eXiTlRoLrHuF;ll,nFrD;!ny;aHeFiend1;eFq1Psh,tfC;!zi5;il,nt3Q;aJolIrFul; keeps,eGge13mFtuna2;al,er;g0Gign;ish;my;aFue3;g4Et;erceHnFrm,t;aFe,i2;l,n32;! ;d Ieb6mGrF;ti6;a6ini7;le;up;bl0iJke,l9n1ZrGst,t,uFvori2;lVx;! F;oFreac2P;ff;nt,r;aTerie,fficie3lSmPnNre3QssLthere4vJxF;a3Pcess,pe2ZtrGuF;be3Wl0W;aFe2C;!va2M;eFil;n,ryday;enF;ti4; Fti8;rou2sui2;iGpF;ty;ne3;abo38dXe2Ci2;gArKsF;tHyF;!gF;oi5;!eF;rn;ly,th1;a00eSiKoHrGuF;e,ll,mb;eaDunk,y;miGne,ubF;le,tfC;na3;dact2Afficu2NrLsHvF;er9i7;ne;creHeas0gruntl0hone2DordGtF;a3ress0;er1;et;e,ty; LaIep,f0Ngene2MliGn9rang0spe2Mtail0voF;ut;be2Kca2ghF;tfC;dFf,r;!ly,pF;an;facto;i1mp,ngero0OppArk,s1E;a05erta04h00ivZlWoJrGuF;dd1r1te;aGisp,owd0uF;ci4el;mp0v0Wz0;ar9lPmMnHoGrrFst1ve1Mward1zy;e2Bu2F;i5l;cre2grIsHtF;emFraD;po18;ta3;ue3;mGpleF;te,x;er11on;d,oF;ni4ss4;eFouB;aFvA;n,r;il;ar18eGiF;ef,l1;ap,eF;sy;in;gey,lmGrF;efC;!i5;a04eZiYlUoNrHuF;r1sy;ly;aJiGoF;ad,wn;ef,gGllFsk;ia3;ht;nd new,sh,ve;gKld,na fi07rJttIuF;nGrgeoF;is;cy;om;ed,i5;us;aGiMoFue;nd,oB;ck,nF;d,k;g,ttAzar8;autifChiHlov0nGst,ttAyoH;er;eYt;nd;ul;ckHd,ld,nkru11rrGwB;dy;en;!wards; priori,b0Qc0Nd0DfraEg08h07l02ma09nZpVrQsOttracti0AutheNvKwF;a8eGkF;wa0X;soF;me;re;ant garGeraF;ge;de;ntW;leep,tonisF;hi5;ab,bitIroHtiF;fiF;ci4;ga3;raD;pFt;are3etiRrF;oprF;ia2;gDtiquF;at0;ry;arIcohGeFiMoof;rt;olF;ic;mi5;ead;ainHgressiGoniF;zi5;ve;st;id; MeKuJvF;aGer9;se;nc0;ed;lt;pt,qF;ua2;hoc,infinitF;um;cuGtu4u2;al;ra2;erPlOoMruLsGunF;da3;e3oGtraFuL;ct;lu2rbi5;ng;te;pt;aFve;rd;aze,e;ra3;nt","Plural":"ad homineTbNcKdGeEfCgBh9kitMlunchQn8othersQp6rooTs4t0us dollaJweekDyesterd2;h2ics,o0ragediP;d0morrows,nighM;ays;erMinO;kiLtu0uper bowls,ysteO;d9ffs;a0robleMurposJ;rGt7;othIumbeA;ead starEo0;mFusF;lacie7o9rand slaH;ri0undE;en7;conomiBgDner3v0xamplB;en8;eaths,inne2o0;g0o1;gi7;rs;anari5eil6i1ottag5redit car0;ds;ti3;ank3o1reakfas0;ts;di0ttl0;es;in0;gs;ms","MaleName":"0:A9;1:9J;2:A0;3:9R;4:93;5:7V;6:9X;7:9B;8:8K;9:7H;A:75;a96b8Kc7Sd6Ye6Af5Vg5Hh4Yi4Oj3Rk3Jl33m25n1Wo1Rp1Iqu1Hr0Xs0EtYusm0vVwLxavi3yDzB;aBor0;cha53h1E;ass2i,oDuB;sEuB;ma,to;nEsDusB;oBsC;uf;ef;at0g;aIeHiCoB;lfga05odrow;lBn16;bDfr9JlBs1;a8GiB;am2Qe,s;eAur;i,nde7Zsl8;de,lBrr6y7;la5t3;an5ern1iB;cBha0nce2Wrg7Sva0;ente,t4J;aPeKhJimIoErCyB;!l3ro7s1;av6OeBoy;nt,v4F;bDdd,mBny;!as,mBoharu;a94ie,y;i9y;!my,othy;eodo0Nia6Aom9;dErB;en5rB;an5eBy;ll,n5;!dy;ic84req,ts3Nyl43;aNcottMeLhIiHoFpenc3tBur1Fylve76zym1;anDeBuaA;f0ph8PrliBve4IwaA;ng;!islaw,l8;lom1uB;leyma7ta;dn8m1;aCeB;ld1rm0;h02ne,qu0Hun,wn;an,basti0k1Nl3Irg3Hth;!y;lEmDntBq3Zul;iBos;a5Ono;!m7Ju4;ik,vaB;d3KtoY;aQeMicKoEuCyB;an,ou;b6dBf67ssel5X;ol2Fy;an,bFcky,dEel,geDh0land8Fm0n5Dosevelt,ry,sCyB;!ce;coe,s;l32r;e43g3n8o8Hri5C;b7Ie89;ar4Xc4Wha6YkB;!ey,y;gCub6x,yBza;ansh,nal4U;g7DiB;na79s;chDfa4l22mCndBpha4ul,y58;al5Iol21;i7Zon;id;ent2int1;aIeEhilDierCol,reB;st1;re;!ip,lip;d7SrDtB;ar,eB;!r;cy,ry;bLt3Iul;liv3m7LrDsCtBum79w6;is,to;ama,c77;i,l3NvB;il4H;athanIeHiDoB;aBel,l0ma0r2H;h,m;cDiCkB;h5Oola;lo;hol9k,ol9;al,d,il,ls1;!i4;aUeSiKoFuByr1;hamDrCstaB;fa,pha;ad,ray;ed,mF;dibo,e,hamDntCrr4EsBussa;es,he;e,y;ad,ed,mB;ad,ed;cFgu4kDlCnBtche5C;a5Yik;an,os,t1;e,olB;aj;ah,hBk8;a4eB;al,l;hBlv2r3P;di,met;ck,hLlKmMnu4rGs1tCuri5xB;!imilian6Q;eo,hCi9tB;!eo,hew,ia;eBis;us,w;cDio,ko,lCsha4WtBv2;i22y;in,on;!el,oIus;colm,ik;amBdi,moud;adB;ou;aMeJiIl2AoEuBy39;c9is,kBth3;aBe;!s;g0nn5HrenDuBwe4K;!iB;e,s;!zo;am,on4;evi,i,la3Yn5KoBroy,st3vi,w3C;!nB;!a4X;mCn5r10uBwB;ren5;ar,oB;nt;aGeChaled,irBrist40uAy2T;k,ollos;i0Wlv2nBrmit,v2;!dCnBt;e0Uy;a43ri3T;na50rBthem;im,l;aZeRiPoDuB;an,liBni0Ost2;an,o,us;aqu2eKhnJnGrEsB;eChB;!ua;!ph;dBge;an,i;!aB;s,thB;an,on;!ath0n4A;!l,sBy;ph;an,e,mB;!m46;ffGrDsB;sBus;!e;a4AemCmai7oBry;me,ni0H;i5Iy;!e01rB;ey,y;cGd6kFmErDsCvi3yB;!d6;on,p3;ed,r1F;al,es;e,ob,ub;kBob;!s1;an,brahJchika,gHk3lija,nuGrEsDtBv0;ai,sB;uki;aac,ha0ma4;a,vinB;!g;k,nngu3W;nacBor;io;im;aKeFina3RoDuByd42;beAgBmber3FsD;h,o;m3ra5sBwa34;se2;aEctDitDnCrB;beAm0;ry;or;th;bIlHmza,ns,o,rCsBya36;an,s0;lEo3BrDuBv8;hi33ki,tB;a,o;is1y;an,ey;!im;ib;aLeIilbe3YlenHord1rDuB;illerBstavo;mo;aDegBov3;!g,orB;io,y;dy,h43nt;!n;ne,oCraB;ld,rd3Q;ffr8rge;bri4rB;la1HrBy;eYy;aOeLiJlIorr0CrB;anDedB;!d2GeBri1K;ri1J;cCkB;!ie,l2;esco,isB;!co,zek;oyd;d4lB;ip;liCng,rnB;anX;pe,x;bi0di;arWdRfra2it0lNmGnFrCsteb0th0uge7vBym6;an,ereH;gi,iCnBv2w2;est33ie;c02k;rique,zo;aGiDmB;aFeB;tt;lCrB;!h0;!io;nu4;beAd1iDliCm3t1v2woB;od;ot1Bs;!as,j35;!d1Xg29mEuCwB;a1Din;arB;do;o0Fu0F;l,nB;est;aSeKieJoDrag0uCwByl0;ay7ight;a7st2;minEnDugCyB;le;!l9;!a1Hn1K;go,icB;!k;go;an,j0lbeAmetriYnFrEsDvCwBxt3;ay7ey;en,in;moZ;ek,ri05;is,nB;is;rt;lKmJnIrDvB;e,iB;!d;iEne08rBw2yl;eBin,yl;lBn;!l;n,us;!e,i4ny;i1Gon;e,l9;as;aXeVhOlFoCraig,urtB;!is;dy,l2nrad,rB;ey,neliBy;us;aEevelaDiByG;fBnt;fo06t1;nd;rDuCyB;!t1;de;en5k;ce;aFeErisCuB;ck;!tB;i0oph3;st3;d,rlBse;es,ie;cBdric,s0N;il;lEmer1rB;ey,lCroBt3;ll;!os,t1;eb,v2;arVePilOlaNobMrCuByr1;ddy,rt1;aGeDi0uCyB;anDce,on;ce,no;nCtB;!t;d0t;dBnd1;!foCl8y;ey;rd;!by;i7ke;al,lF;nDrBshoi;at,naBt;rd0F;!iCjam2nB;ie,y;to;ry,t;ar0Qb0Id0Fgu0Dhme0Cid6jani,lVmTnLputsiKrCsaBu0Dya0ziz;hi;aHchGi4jun,maEnCon,tBy0;hur,u05;av,oB;ld;an,nd04;el;ie;ta;aq;dGgelZtB;hoEoB;i7nB;!iW;ne;ny;reBy;!s,w;ir,mBos;ar;!an,beOeIfFi,lEonDt1vB;aMin;on;so,zo;an,en;onCrB;edJ;so;jEksandDssExB;!and3is;er;ar,er;andB;ro;rtB;!o;en;d,t;st2;in;amCoBri0vik;lfo;!a;dDel,rahCuB;!bakr,lfazl;am;allEel,oulaye,ulB;lCrahm0;an;ah,o;ah;av,on","Expression":"aZbYdTeRfuck,gQhKlHmGnFoCpAsh9u7voi01w3y0;a1eKu0;ck,p;!a,hoo,y;h1ow,t0;af,f;e0oa;e,w;gh,h0;! huh,-Oh,m;eesh,hh,it;ff,hew,l0sst;ease,z;h1o0w,y;h,o,ps;!h;ah,ope;eh,mm;m1ol0;!s;ao,fao;a3e1i,mm,urr0;ah;e,ll0y;!o;ha0i;!ha;ah,ee,oodbye,rr;e0h,t cetera,ww;k,p;'3a0uh;m0ng;mit,n0;!it;oh;ah,oo,ye; 1h0rgh;!em;la","Unit":"0:1C;a17b15c0Rd0Qe0Of0Kg0Gh0Din0Cjoule0k06l04mSnRoQpLqKrJsDt8volt0w7y5z4°2µ1;g,s;c,f,n,r1;a,e;b,e2;a0Pb,d0Vo1;tta0J;att0b,ebers;able4b3e2on1sp;!ne0;a2ra0F;!l,sp;spo07;eQiemens0q1; ft,uare 1;c0Jd0If3i0Gkilo0Km1ya0F;e0Nil1;e0li0I;eet0o0E;ankine0eaumur0;t,uart0;a4e2i1ou0Pt;c0Mnt0;rcent,t1;a01roleum barrels;!scal0;hm0uWz;an0HewtS;/s,b,e7g,i3l,m2p1s,²,³;h,s;!²;!/h,cro3l1n;e1li06;! EsD²;g06s0B;gaQter1;! 2s1;! 1;per se1;cond;b,iZm,u1x;men0x0;b,elvins,g,ilo2m1nQ;!/h,ph,²;byYgWmeter1;! 2s1;! 1;per hour;²,³;a,e1g,z;ct1nrys,rtz0;aWogP;al2b,iga9ra1;in0m0;!l1;on0;a3emtOl1oCtG; oz,uid ou1;nce0;hrenheit0rad0;b,xa1;byH;eciCg,l,mA;arat0eAg,l,m9oulomb0u1;bic 1p0;c5d4fo3i2meAya1;rd0;nch0;ot0;eci2;enti1;me4;!²,³;lsius0nti1;g2li1me1;ter0;ram0;bl,y1;te0;c4tt1;os1;eco1;nd0;re0;!s","Place":"aXbWcSdRevergladTfPgNhMiKjfk,kIlGmEneBoAp7que,rd,s6te5u4v3w0yyz;is1y0;!o;!c;a,t;safa,t;nn,x;ask,fo,t,yd;acifSe1h0;l,x;k,nnG;kla,nt,rd;b1w 0;england,hampshire,jersey,mexico;!r;co,i0t,uc;a,nn;a0gw,hr;s,x;an0ul;!s;a0cn,ndianGst;!x;kg,nd,wy;a,reat 0;britain,lak4;co,l0ra;!a;ak,en,fw,ist,xb;al2dg,gk,lt,o1r0t;es;lo,nn;!if;cn,kk,lvd;ms,ntarct2r1tl0ve;!ant1;ct0iz;ic0; ocean","FemaleName":"0:81;1:7E;2:7G;3:7Y;4:65;5:7P;6:7T;7:7O;8:7U;9:7C;A:6K;B:7R;C:6X;D:79;a78b6Pc5Ud5Ce4Rf4Jg47h3Zi3Uj38k2Sl21m1An14o12p0Ur0FsYtNursu9vIwGyEza7;olan2vE;etDon5Z;an2enMhi6PilE;a,la,ma;aHeFiE;ctor1o9rgin1vi3B;l4VrE;a,na,oniA;len5Ones7N;aLeJheIi3onHrE;acCiFuE;dy;c1na,s8;i4Vya;l4Nres0;o3GrE;e1Oi,ri;bit8mEn29ra,s8;a7iEmy;!ka;aTel4HhLiKoItHuFyE;b7Tlv1;e,sEzV;an17i;acCel1H;f1nEph1;d7ia,ja,ya;lv1mon0;aHeEi24;e3i9lFrE;i,yl;ia,ly;nFrEu3w3;i,on;a,ia,nEon;a,on;b24i2l5Ymant8nd7raB;aPeLhon2i5oFuE;by,th;bIch4Pn2sFxE;an4W;aFeE;ma2Ut5;!lind;er5yn;bFnE;a,ee;a,eE;cAkaB;chEmo3qu3I;a3HelEi2;!e,le;aHeGhylFriE;scil0Oyamva2;is,lis;arl,t7;ige,mGrvati,tricFulE;a,etDin0;a,e,ia;!e9;f4BlE;ga,iv1;aIelHiForE;a,ma;cEkki,na;ho2No2N;!l;di6Hi36o0Qtas8;aOeKiHonFrignayani,uri2ZyrE;a,na,t2J;a,iE;ca,q3G;ch3SlFrE;an2iam;dred,iA;ag1DgGliFrE;ced63edi36;n2s5Q;an,han;bSdel4e,gdale3li59nRrHtil2uGvFx4yE;a,ra;is;de,re6;cMgKiGl3Fs8tFyanE;!n;a,ha,i3;aFb2Hja,l2Ena,sEtza;a,ol,sa;!nE;!a,e,n0;arEo,r4AueriD;et4Ai5;elLia;dakran5on,ue9;el,le;aXeSiOoKuGyE;d1nE;!a,da,e4Vn1D;ciGelFiEpe;sa;a,la;a,l3Un2;is,la,rEui2Q;aFeEna,ra4;n0t5;!in0;lGndEsa;a,sE;ay,ey,i,y;a,i0Fli0F;aHiGla,nFoEslCt1M;la,na;a,o7;gh,la;!h,n07;don2Hna,ra,tHurFvern0xE;mi;a,eE;l,n;as8is8oE;nEya;ya;aMeJhadija,iGrE;istEy2G;a,en,in0M;mErst6;!beE;rlC;is8lFnd7rE;i,ri;ey,i,lCy;nyakumari,rItFvi5yE;!la;aFe,hEi3Cri3y;ar4er4le6r12;ri3;a,en,iEla;!ma,n;aTeNilKoGuE;anEdi1Fl1st4;a,i5;!anGcel0VdFhan1Rl3Eni,seEva3y37;fi3ph4;i32y;!a,e,n02;!iFlE;!iE;an;anHle3nFri,sE;iAsiA;a,if3LnE;a,if3K;a,e3Cin0nE;a,e3Bin0;cHde,nEsm4vie7;a,eFiE;ce,n0s;!l2At2G;l0EquelE;in0yn;da,mog2Vngrid,rHsEva;abelFiE;do7;!a,e,l0;en0ma;aIeGilE;aEda,laE;ry;ath33i26lenEnriet5;!a,e;nFrE;i21ri21;aBnaB;aMeKiJlHrFwenE;!dolY;acEetch6;e,ie9;adys,enEor1;a,da,na;na,seH;nevieve,orgi0OrE;ald4trude;brielFil,le,yE;le;a,e,le;aKeIlorHrE;ancEe2ie2;es,iE;n0sA;a,en1V;lErn;ic1;tiPy1P;dWile6k5lPmOrMstJtHuGvE;a,elE;yn;gen1la,ni1O;hEta;el;eEh28;lEr;a,e,l0;iEma,nest4;ca,ka,n;ma;a4eIiFl6ma,oiVsa,vE;a,i7;sEzaF;aEe;!beH;anor,nE;!a;iEna;th;aReKiJoE;lHminiqGnPrE;a,e6is,othE;ea,y;ue;ly,or24;anWna;anJbIe,lGnEsir1Z;a,iE;se;a,ia,la,orE;es,is;oraBra;a,na;m1nFphn0rlE;a,en0;a,iE;el08;aYeVhSlOoHrEynth1;isFyE;stal;ti3;lJnsHrEur07;a,inFnE;el1;a,e,n0;tanEuelo;ce,za;e6le6;aEeo;ire,rFudE;etDia;a,i0A;arl0GeFloe,ristE;a,in0;ls0Qryl;cFlE;esDi1D;el1il0Y;itlin,milMndLrIsHtE;ali3hE;er4le6y;in0;a0Usa0U;a,la,meFolE;!e,in0yn;la,n;aViV;e,le;arbVeMiKlKoni5rE;anIen2iEooke;dgFtE;tnC;etE;!te;di;anA;ca;atriLcky,lin2rItFulaBverE;ly;h,tE;e,yE;!e;nEt8;adOiE;ce;ce,z;a7ra;biga0Kd0Egn0Di08lZmVnIrGshlCudrEva;a,ey,i,y;ey,i,y;lEpi5;en0;a,dNeLgelJiIja,nGtoE;inEn1;etD;!a,eIiE;ka;ka,ta;a,iE;a,ca,n0;!tD;te;je9rE;ea;la;an2bFel1i3y;ia;er;da;ber5exaJiGma,ta,yE;a,sE;a,sa;cFsE;a,ha,on;e,ia;nd7;ra;ta;c8da,le6mFshaB;!h;ee;en;ha;es;a,elGriE;a3en0;na;e,iE;a,n0;a,e;il","FirstName":"aIblair,cGdevFguadalupe,jAk8lashawn,m5r2sh0trinity;ay,e0iloh;a,lby;e1o0;bin,sario;g1ne;ar1el,org0;an;ion,lo;asAe0;ls9nyatta,rry;a1e0;an,ss2;ime,m0n;ie,m0;ie;an,on;as0heyenne;ey,sidy;ndra,ubr0;ey","Month":"aAdec8feb6j2mar,nov8oct1sep0;!t7;!o7;an4u0;l1n0;!e;!y;!r0;uary;!em0;ber;pr1ug0;!ust;!il","Preposition":"'o,-,aKbHcGdFexcept,from,inEmidPnotwithstandiQoCpRqua,sBt6u3vi2w0;/o,hereMith0;!in,oQ;a,s-a-vis;n1p0;!on;like,til;h1ill,o0;!wards;an,r0;ough0u;!oH;ans,ince,o that;',f0n1ut;!f;!to;espite,own,u3;hez,irca;ar1e0y;low,sides,tween;ri6;',bo7cross,ft6lo5m3propos,round,s1t0;!op;! long 0;as;id0ong0;!st;ng;er;ut","Honorific":"aTbScJdIeHgFhon,jr,king,lDmAop,p6queen,r3s0;e1fc,gt,ir,r,u0;ltKpt,rg;cretary,n8;abbi,e0;p6s,v0;!erend;astJhd,r0vt;eside6ince1of0;!essH;!ss;a1essrs,istDlle,me,r0str;!s;j,yD;ady,ieutena0lb,ord,t;nt;en0ov;!eral;sq,xcellency;et,oct7r,utchess;apt7hance5mdr,o0pl;m2ngress0unci4;m0wom0;an;dr,mand0;er;ll0;or;!ain;ldg,rig;d0sst,tty,yatullah;j,m,v","Superlative":"0:18;1:19;a18b12c0Rd0Pe0Mf0Gg0Bh01ic0j00lVmSnQoPpLquiKrIs7t5u3vOw2yel04zan0;et14ind0or1r0;gl0n2;lik00tid0;al0Qhirst0i2oug0F;d0gh10mYn0;aBcar0ex0h8i0Tkinn0l6m5p4t2weE;ee0Ura2;i0Wn0Z;ic0righ0O;al0Ke0P;i2oU;e1gh0Tm0;a3i2;e1n0;d0lPr0M;in0Hlt0vv0;ead0itz0o2uT;om0s0ug01;et0M;hon0l4r2un0;ic2ouP;e0kl0;aine1;il0ran0J;arroGeVice1o2;is0s0;e2o0F;a2lC;sl0t0;a0Bi5o2uck0;ft0n8u3v8w2;e1l0;de1s0;gh07k5v5;o00uic0;a9e6iZo2u08;l3m2t04;el0;lo2;we1;a2ft0;d0v2;enl0ie1;pp0r2;de1;a5h4idd0loom0oof0r2;a0e2isl0oov0;atTie1;asKosK;ie1ud0;a6itQl5oam0rienHu2;lEnn0r2;r0t2;he1;atM;nc0tLult0;a2mpt0;r2s0;l0thl0;amDe2irt0r0ul6;a7eC;age0h8loud0o5r3u2;d5rl0;is9ue2;le1;s3war2z0;dl0;tl0;e3i2;ll0;a2es0;pe1;awd0ig6lo5ounc0ri3u2;rl0s0;gh2;te1;od0;ge1;ngr0;ie1;st","Comparative":"0:16;1:0Z;a15b0Yc0Sd0Oe0Lf0Fg08h02i13j01lWmTnRoQpMquie0JrIs7t5u3vPw2yell04za0H;e0Iin13orse,ri0;g1n2;like1ti11;a0Lhirs0Mi2ouG;di0g0Wme1ni0;aBca10exi0h8il1kin0Bl6m5p4t2wee0D;eep0ra2;ig0Tng0;i0Uright1;a0Fel1;i2ow0;er,g0Pmi0;a3i2;er,ni0;di0llP;dd0int1l0AvU;e4itzi0o2ud0;o00si0u2;gh0;a0Kdd0;hoWl4r2uW;ic2oud0;ei0k1;ain0;i1ld0rang0;arrEeaTo2;i06si0;e2oo0C;a2llB;s1ti0;i5o2ucki0;fUne1u3ve1w2;er,li0;d0si0;g02ke1ve1;ol1ui03;appi0e4il1o2u04;ll2me1tt0;ow0;a2fM;di0v2;en1i0;a7h5idXlooBoofi0r2;ay0e3is1oo2;vi0;a9en0y0;a2o2;st1;i0uR;a6i5la5oa4riend1u2;ll0n2rR;ni0;mi0;tt0;nKtt0ul7;a2mp6;r2si0;li0th1;e4ir3ri0u2;ll0;ti0;ad1ep0;agei0h4louDo3ruel0u2;dd1r1;ld0st1ward1zi0;e2il1;ap0e2;si0;aw7igg0loo7o5rig4u2;r1si0;li0;ht0;ld0un2;ci0;di0;ng2;ri0;er","Gerund":"0:73;1:71;2:6N;3:72;4:70;5:69;6:6F;7:5T;8:5C;9:6X;A:6R;B:6Y;C:5W;D:66;E:6T;F:6L;a6Fb5Zc54d4Ce3Wf3Kg3Dh37i2Wj2Uk2Rl2Im27n24o20p1Iquest46r0Ws04tVuRvOwHyG;aw3ielA;aJe20hist5iIoGrec7;n26rG;ki0rC;n3pi0s6t60;l7nHrGs6tc6vi0;ni0ran1;de2ti0;aHiGo1;e6UsE;l6FrC;nGplif1r9si0;dGi1l5U;erGoi0;g3Wmi3;aNeMhKoJrHuGw1I;ni0r3;aGeat1ic7us1yi0;di0i3ns49p8ve5;r31uc6;an7rG;eate3iB;nAs1;ki0l7rg5Es1;a06c03e00hXiAkip8lWmVnTor1pRtKuGwi05;bmEc7f42gg0XppHrGsp41;gi0pri4rounAviB;lCoG;r1si0;aLeKoJrHuG;dCmb5;ai3eGiB;ngthe3s4tc6;p8ri0;mDp8;i3ki0r0Tyi0;it1las6oi5rinG;gi0k5;eGo61;a7e36;as6el5i5o7;amDee8ip8;aHoG;c7ut1;ri0vi0;arc6eHl3HnGrBtt2J;di0si0;ki0mi0;a5Jo2rG;aGe5Q;tc6;i5vi0;aZeIis7oHuG;b4Oi3li0s6;aDl5;aVcUduFfTgSj37lQmPn1pNquMsJtIvGwarA;eGie5J;a5li0r4;ai3i2ur3;eHis1or1tG;i0r02;mb5n1rB;es1i2;ai2lGor1resY;aFyi0;ai3embe2inAoB;aGen1iByi0;ti0xi0;arAr45;l2Tu4;al5eiBko3ogni2B;c6li2Aso3;i4nG;gi0ki0;aWerUiTlRoPrIuG;bl2Hl5m8rGs6t1;cha4s4L;ai4eJin1oG;cHduFgr53hibEmi4tGvi2C;es1;es4laiD;dHpa2s4vG;en1;ic1;in1p8rt3XsGu2;i0ti0;a3MuG;g9n9;c7li0;forDmEsG;is1;c7in1rt4Es4u4yi0;bIccu2f27pera1utweHverGwi0;coDlap8ta7u4whelD;ig6;serBt2Y;eHotiG;ci0ng;eAgl20;aMeLiJoIuG;ltiplCrG;de2;nito2vi0;nGrro2;di0g5u4;asu2l1nt18r9ssa9;int2Nke d3Rna9rHtG;c6te2;kGrC;et1i0;aMeLiIoGum8;aGc7o7vi0;di0ni0;f1ki0mEnHstGvi0;e3i0;i0ki0;ar3c0Ct1;bo2c7uG;g6nc6;ee8iHnoG;c7wi0;c7l5s4;amDoi3uG;d9m8;gno2mOnGss39;cMdica1flu0IhLsJterIvG;eGolB;n1s1;ac1fe2;is1tG;al5r1V;abEibE;luAr1C;agi3pG;or1roB;aKeIi2oHuG;n1rrC;ld fUno2pi0s1;aGl8;di0t1;nd5ppe3rDti0;aLet1los4oKrIuG;arGes4iA;ant35di0;aGip8o33umb5;b24s8;e4v1G;i3t21zi0;aQeaPiNlLoHracGunct01;tu2;cu4lIrG;ci0ecGmi0s2X;as1;di0lo2U;as6oG;at1oAwi0;gu2lGn08xi0;i0li0;ri0tu2;i5l5vo2;ar3chUdEl0AmRnNstMvalua1xG;a0CcKerJis1pGtenA;anAeGl13or1r2R;c1riG;enF;ci4ti0;e5han9;ablXima1;ab5couZdIga9hanFjoCric6su2te2vG;isGyi0;io3;i0u2;bodCer9pG;hasiGo25;zi0;oi0;a06eWiKoJrGum8yi0;aHeGif1op8;aDs4;gi0i3;d9i0ub1;fQp8rPsIvG;iGorF;di0ng;aKcHli7mis4plaCtinguG;is6;oGus4;uGve2;ra9;gr1VppG;ea2roB;ec1;fe2;cMeDfKlive2m09nCpJsHterG;mi3;cri0QerBiGtroC;g3ri0;ar1enAic1osE;eGi3yi0;at1c1;iAlHrG;ea4;a2i3;ma9nFri0ti0;a08e07h04i1l00oJrGut1;aHeGos4us6yi0;a1dE;c7vi0;llVmRnHo7pCuGve2;g6n1;cOdNfMsKtIvG;eGinF;r1yi0;ai3in0SrG;as1ibu1o5;iGtitu1;de2s1;irDorDron1;uc1;er3;bi3for1i0mIpG;a2lG;ai3;anA;ap4ec1;aIiGo4;n9p8;pi0;iDwi0;aHeG;at1c7wi0;llen9n9r9;a4nte2;l5nce5rGte2u4;i0rC;aTePlNoKrJuG;dgHrG;ni0yi0;et1;a7us6;as1mHos1rro0FtGunF;he2;bi0;oc7;ki0;a2coDlInefEtG;raC;yi0;ieBon9;nGsi0;gi0ni0;bando3c0Bd07fforAgr05iAl01mu4nXppSrPsJttGvoiAwaE;acHeGrac1;mp1nA;hi0ki0;ki0sG;eIiHocia1uD;mi0;g3s1;mb5r1;li0;chiHgGriB;ui0;ng,vi0;eaJlIroG;ac6vi0;hi0;auAyi0;li0ri0;nounFsHtG;icipa1;we2;ci0;lHte2;ri0;e9oG;wi0;ei0;di0;dHjus1mEop1;it1;i0rG;es4;cIhieBknowled9tiva1;gi0;vi0;ep1oun1u4;si0;ti0;ni0;ng","Determiner":"aAboth,d8e5few,l3mu7neiCown,plenty,some,th2various,wh0;at0ich0;evB;at,e3is,ose;a,e0;!ast,s;a1i6l0nough,very;!se;ch;e0u;!s;!n0;!o0y;th0;er","PastTense":"0:7I;1:72;2:78;3:6G;4:5L;5:77;a6Jb62c55d49e3Tf3Gg39h32i2Qj2Nk2Jl2Am20n1Xo1Tp1Cques3Hr0PsXtNuJvGw7y6;awn0ield0;aBeAhistl0i9o6re69;n7r6;k0ri0;!d1;p0sh0th1;i1Ont;lk0n7r6sh0t6Qv0;n0ra7B;d1t0;a7i6ot0;ew0s5;lu0ri0;n6pli3Xrg0s0;d6it0lo5W;er6o0;go0m4;aEeDhBoAr7u6wi2;n0rn0;a7eat,i6u2;ck0ed;d0in0nsf1p,v5F;ok,rt3Iu6B;ank0r6;eat2Jiv0;nt,st0;lk0rg5Est0;aWcUeQhNkip,lMmLneKoJpHtAu6wit66;bm5ck0ff1gge2pp7r6spe3;g0pr34rou4Uv60;li0o6;rt0s0;aBeAo9r7u6;di0m5U;a4e6iv0;ngth28ss0t5Y;ok,pp0r0;mm0p;in0r0Jy0;it,la53oil0rin6;g0kl0;de,rt0;ak0ez0w;a4Ze1Pil0ok0;a1Ge1Dip;a7o6;ck0ut;r0v0;ar5Le8le3n7rv0t6;!tl0;s0t;k0m0;a5Kor0r6;at5Gew0;il0v0;aQe8isk0o7u6;bb0in0l0sh0;am0ll0;aMcLdVfKgJje3lGmFnt0pDquCs9t8v6wa5N;e6iew0;al0l0rs0;a4ir0u3V;e7i2o3Yt6;ed,ri3;m4Znt0rv0;e2ir0;air0l6o3Urese5P;ac0i0;a4emb1i3Rov0;a7e5Mi6;ed,v0;t0x0;a59r44;le3us0;a0Ue4Rk5Jogn25;ch0l24s5I;is0n6;g0k0;aLerKiJlGoEr8u6;b1Tll0mp0r6sh0t;ch2Rsu0;a1OeAi5Ao6;c8d7gre54hib5m1Nte2v6;ed,o2P;uc0;e51laim0;di3p29ss0ve55;i54pp0rtr3Xs6ur0;ed,t0;a7u6;gg0ng0;n0y0;ck0l0;fo2Ym5si2;ck0i4Xrt4Gss0us0;b8cc1Hff1per4Tutwei7ver6w0;c3Rlap,to3Jus0whelm0;gh0;se22ta4;e7ot6;ed,ic0;ed0gle3;aBeAi8o7u6;ltipli0rd1;nit0Vv0;n6rr0U;d0gl0us0;as15lt0n0Hrg0ss24;inta4ke du0n23r7t6;ch0t1;ke6ri0;d,t;aCeBi9o7u6;mp0nk;a6ck0ok0v0;d0n0;ft0k0m5n0ste6v0;d,n0;a2Act0Ut;b0Gck0u6;gh0n3L;e8i7n6;ew,o2W;ck0ll0ss0;pt;a7o4u6;dg0mp0;mm0;gn07mFn6ssu0;cDdic3Vflue3IhCs9ter8v6;e6olv0;nt0st0;a3f1;i2t6;a6ru3;ll0;ab5ib5;l36re16;ag4p6;o1Tr0R;aAe8ir0o7u6;nt0rri0;ld fa2nTp0st0;a6lp0;d0t;ndl0pp6rm0t0;en0;aBlo3BoAr8u6;ar6e3Ao0W;ante0d0;a6ew,ip,um2L;bb0sp0;t,ve1C;in0th1z0;aGeEiClAo7ractUunc6;ti3A;c2Sl7r6;c0eca2m0se0;d0l1V;a1Rew,o6;at,od0;gNl6ni1Px0;ed,l0;a6ll;r0tK;il0v6;or0;a0Xcho0d5le3mInEstCvalu2Tx6;am4cAer8i2p6te06;a0Ye6la4o0Zre2P;ct0rie2D;c6t0;is0;el0ha1L;ab6im2M;li1B;a1Vcour07d8g07ha27jUri21s8t7vi6;ed,si2O;!er0;ur0;bodi0erg0p6;has6ow1;iz0;aZeLiBoAr6ump0;a8e7i6opp0;ft0;am0ss0;g0in0;dg0ubt0;d,ed,ff1p,re3s8v6;ed,o6;de,rc0;a8c6lik0mi22pl10tingui0U;o6u21;urPv1;g1Tpp6;e7r6;ov0;ar0;cFem0fCliv1ma04ni0pAs6term4;crib0e8i7tr6;oy0;gn0r0;rv0;a00e6i3os5;nt;e7i6;ed,n0;at,ct0;l8o7re6;as0;de;ar0in0;m6nc0r0t0;ag0;aZeYhVit0lSo9r6ut;a7e6i0o1Eu06;at0d5;ck0v0;llNmIn7ok0pi0u6v1;gh0nt0;ceFdu3fDsAt8v6;e6i0X;rt0y0;a4inu0r6;a2ib7ol0;i7tit6;ut0;d1st0;i6o6ro16;rm0;rn0;b4fo8ma7p6;ar0la4;nd0;rt0;in0;aps0e3;a7i6os0;ng0p;im0w0;a7e6;at,ck0w0;lleNng0rg0;as0nt1;ll0me,nc7r6t1us0;ed,ri0;el0;aKeFloEoCr9u6;dg7r6;i0n0;et;o7u6;sh0;ok;a2mb0o2rr6th1u00;ew;ck0;ar0c9l7nef5tr6;ay0;i09o6;ng0;ame;n6s0;g0n0;band08c02dXffoWgVid0lSmRnOppIrFsAtt6void0wa5;ac8e7ra3;ct0;m04nt;h0k0;k0s6;e7i6ociXum0;gn0st0;m6rt0;bl0;ch7gu0r6;iv0;ed,iv0;eaAl8ro6;a6v0;ch0;a6i0;ud0;l0r0;nou7sw1t6;icipJ;nc0;us0;le7t1;er0;g0w;re0;rd0;d8ju2m5oF;it;st0;ed,re6;ss0;c8hi7knowledg0tiv6;at0;ev0;e7ou6us0;nt0;pt0;on0;ed","Adverb":"a06by 04d00eXfShQinPjustOkinda,mMnKoFpDquite,rAs6t3up2very,w1ye0;p,s;ay,ell; to,wards5;h1o0wiN;o,t6ward;en,us;everal,o0uch;!me1on,rt0; of;hVtimes,w05;a1e0;alQ;ndomPthL;ar excellCer0oint blank; Khaps;f3n0;ce0ly;! 0;agYmoS; courFten;ewHo0; longCt withstanding;aybe,eanwhi9ore0;!ovA;! aboR;deed,steS;en0;ce;or1urther0;!moH; 0ev3;examp0good,suF;le;n mas1v0;er;se;e0irect1; 1finite0;ly;ju7trop;far,n0;ow; CbroBd nauseam,gAl5ny2part,side,t 0w3;be5l0mo5wor5;arge,ea4;mo1w0;ay;re;l 1mo0one,ready,so,ways;st;b1t0;hat;ut;ain;ad;lot,posteriori","Actor":"at8br7d4e3gu6k2linkAneedAovert7prov6s1t0;ak9e8;e7id8l0t5;eep7;nd6x4;e0iv1;c0pe3;id3;ak2;te0;nd0;er","Participle":"a01bZcUdPeJfHgFh07iDkno06lCmAoper07p6r4s1tortNunder0;goFmW;n1ought,p0urprJ;at,rV;o02uck;aGe0;alIcognIl01;a2r0;aDo0;mCvZ;id,rtT;anIe0inglX;asBssH;e5inV;magJn0;dicSflueP;o0roQ;ne;allQeat5ig5loOolNra0;ct4;mphas4n2stimNvaluNx0;amCerc0perieJ;isM;c3d0g6haHs0;urK;izJ;a2e1isc0reamt;our2;cl5f5term5;m0ncFtF;agE;l3o0reC;m0nvi8;b0e;inA;ung;ecome,o0rok8;rro5u3;ctiv5l3n0ssoci5;nou1t0;icip3;nc3;lo0;wn;at0;en","Demonym":"0:17;a0Wb0Mc0Bd09e08f06g03h01iXjUkSlOmKnHomGpCqatari,rAs6t4u3v2wel0Qz1;am0Eimbabwe0;enezuel0ietnam0G;g8krai11;aiwShai,rinida0Hu1;ni0Prkmen;a3cot0Je2ingapoNlovak,oma0Tpa04udQw1y0X;edi0Jiss;negal0Ar07;mo0uT;o5us0Kw1;and0;a2eru0Ghilipp0Po1;li0Drtugu05;kist2lesti0Qna1raguay0;ma0P;ani;amiYi1orweO;caragu0geri1;an,en;a2ex0Mo1;ngo0Erocc0;cedo0Ila1;gasy,y07;a3eb8i1;b1thua0F;e0Dy0;o,t01;azakh,eny0o1uwaiti;re0;a1orda0A;ma0Bp1;anM;celandic,nd3r1sraeli,ta02vo06;a1iS;ni0qi;i0oneU;aiCin1ondur0unM;di;amCe1hanai0reek,uatemal0;or1rm0;gi0;i1ren6;lipino,n3;cuadoVgyp5ngliIstoWthiopi0urope0;a1ominXut3;niG;a8h5o3roa2ub0ze1;ch;ti0;lom1ngol4;bi0;a5i1;le0n1;ese;liforLm1na2;bo1erooK;di0;a9el7o5r2ul1;gaG;aziBi1;ti1;sh;li1sD;vi0;aru1gi0;si0;ngladeshi,sque;f9l6merAngol0r4si0us1;sie,tr1;a1i0;li0;gent1me4;ine;ba2ge1;ri0;ni0;gh0r1;ic0;an","Doer":"aKbHcGdDengineJfBgardenJhAinstructQjournalMlawyJm9nurse,o8p5r3s1t0;echnDherapL;ailOcientKoldiHu0;pervLrgeon;e0oofF;ceptionHsearD;hotographDlumbDoli1r0sychologG;actitionCogrammC;cem7t6;fficApeG;echanic,inist9us4;airdress8ousekeep8;arm7ire0;fight6m2;eputy,iet0;ici0;an;arpent2lerk;ricklay1ut0;ch0;er;ccoun6d2ge7r0ssis6ttenda7;chitect,t0;ist;minist1v0;is1;rat0;or;ta0;nt","SportsTeam":"0:1M;1:1T;2:1U;a1Rb1Dc0Zd0Qfc dallas,g0Nhouston 0Mindiana0Ljacksonville jagua0k0Il0Fm02newVoRpKqueens parkJrIsAt5utah jazz,vancouver whitecaps,w3yY;ashington 3est ham0Xh16;natio21redski1wizar12;ampa bay 6e5o3;ronto 3ttenham hotspur;blu1Hrapto0;nnessee tita1xasD;buccanee0ra1G;a7eattle 5heffield0Qporting kansas13t3;. louis 3oke12;c1Srams;mari02s3;eah1IounI;cramento Sn 3;antonio spu0diego 3francisco gi0Bjose earthquak2;char0EpaB;eal salt lake,o04; ran0C;a8h5ittsburgh 4ortland t3;imbe0rail blaze0;pirat2steele0;il3oenix su1;adelphia 3li2;eagl2philNunE;dr2;akland 4klahoma city thunder,r3;i10lando magic;athle0Trai3;de0; 3castle05;england 6orleans 5york 3;city fc,giUje0Lkn02me0Lred bul19y3;anke2;pelica1sain0J;patrio0Irevolut3;ion;aBe9i3ontreal impact;ami 7lwaukee b6nnesota 3;t4u0Rvi3;kings;imberwolv2wi1;re0Cuc0W;dolphi1heat,marli1;mphis grizz3ts;li2;nchester 5r3vN;i3li1;ne0;c00u0H;a4eicesterYos angeles 3;clippe0dodFlaA; galaxy,ke0;ansas city 3nH;chiefs,ro3;ya0M; pace0polis colX;astr0Edynamo,rockeWtexa1;i4olden state warrio0reen bay pac3;ke0;anT;.c.Aallas 7e3i0Cod5;nver 5troit 3;lio1pisto1ti3;ge0;bronc06nuggeO;cowboUmav3;er3;ic06; uX;arCelNh8incinnati 6leveland 5ol3;orado r3umbus crew sc;api5ocki2;brow1cavalie0india1;benga03re3;ds;arlotte horCicago 3;b4cubs,fire,wh3;iteE;ea0ulY;di3olina panthe0;ff3naW; c3;ity;altimore ElAoston 7r3uffalo bilT;av2e5ooklyn 3;ne3;ts;we0;cel4red3; sox;tics;ackburn rove0u3;e ja3;ys;rs;ori3rave1;ol2;rizona Ast8tlanta 3;brav2falco1h4u3;nited;aw9;ns;es;on villa,r3;os;c5di3;amondbac3;ks;ardi3;na3;ls","Country":"0:38;1:2L;a2Wb2Dc21d1Xe1Rf1Lg1Bh19i13j11k0Zl0Um0Gn05om3CpZqat1JrXsKtCu6v4wal3yemTz2;a24imbabwe;es,lis and futu2X;a2enezue31ietnam;nuatu,tican city;.5gTkraiZnited 3ruXs2zbeE;a,sr;arab emirat0Kkingdom,states2;! of am2X;k.,s.2; 27a.;a7haBimor-les0Bo6rinidad4u2;nis0rk2valu;ey,me2Xs and caic1T; and 2-2;toba1J;go,kel0Ynga;iw2Vji2nz2R;ki2T;aCcotl1eBi8lov7o5pa2Bri lanka,u4w2yr0;az2ed9itzerl1;il1;d2Qriname;lomon1Vmal0uth 2;afr2IkLsud2O;ak0en0;erra leoEn2;gapo1Wt maart2;en;negKrb0ychellY;int 2moa,n marino,udi arab0;hele24luc0mart1Z;epublic of ir0Com2Cuss0w2;an25;a3eHhilippinTitcairn1Ko2uerto riM;l1rtugE;ki2Bl3nama,pua new0Tra2;gu6;au,esti2;ne;aAe8i6or2;folk1Gth3w2;ay; k2ern mariana1B;or0M;caragua,ger2ue;!ia;p2ther18w zeal1;al;mib0u2;ru;a6exi5icro09o2yanm04;ldova,n2roc4zamb9;a3gol0t2;enegro,serrat;co;c9dagascZl6r4urit3yot2;te;an0i14;shall0Vtin2;ique;a3div2i,ta;es;wi,ys0;ao,ed00;a5e4i2uxembourg;b2echtenste10thu1E;er0ya;ban0Gsotho;os,tv0;azakh1De2iriba02osovo,uwait,yrgyz1D;eling0Jnya;a2erF;ma15p1B;c6nd5r3s2taly,vory coast;le of m19rael;a2el1;n,q;ia,oI;el1;aiSon2ungary;dur0Mg kong;aAermany,ha0Pibralt9re7u2;a5ern4inea2ya0O;!-biss2;au;sey;deloupe,m,tema0P;e2na0M;ce,nl1;ar;bTmb0;a6i5r2;ance,ench 2;guia0Dpoly2;nes0;ji,nl1;lklandTroeT;ast tim6cu5gypt,l salv5ngl1quatorial3ritr4st2thiop0;on0; guin2;ea;ad2;or;enmark,jibou4ominica3r con2;go;!n B;ti;aAentral african 9h7o4roat0u3yprQzech2; 8ia;ba,racao;c3lo2morPngo-brazzaville,okFsta r03te d'ivoiK;mb0;osD;i2ristmasF;le,na;republic;m2naTpe verde,yman9;bod0ero2;on;aFeChut00o8r4u2;lgar0r2;kina faso,ma,undi;azil,itish 2unei;virgin2; is2;lands;liv0nai4snia and herzegoviGtswaGuvet2; isl1;and;re;l2n7rmuF;ar2gium,ize;us;h3ngladesh,rbad2;os;am3ra2;in;as;fghaFlCmAn5r3ustr2zerbaijH;al0ia;genti2men0uba;na;dorra,g4t2;arct6igua and barbu2;da;o2uil2;la;er2;ica;b2ger0;an0;ia;ni2;st2;an","Pronoun":"'em,elle,h4i3me,ourselves,she5th1us,we,you0;!rself;e0ou;m,y;!l,t;e0im;!'s","Possessive":"anyAh5its,m3noCo1sometBthe0yo1;ir1mselves;ur0;!s;i8y0;!se4;er1i0;mse2s;!s0;!e0;lf;o1t0;hing;ne","City":"a2Pb1Zc1Qd1Le1Kf1Ig1Dh15i13jakar2Ak0Ul0Om0Bn07o05pUquiTrQsJtAu9v6w3y1z0;agreb,uri1T;ang1Ne0okohama;katerin1Crev2X;ars1ellingt1Lin0rocl1;nipeg,terth0S;aw;a1i0;en2Clni2U;lenc2Pncouv0Dr2B;lan bat0Atrecht;a6bilisi,e5he4i3o2rondheim,u0;nSr0;in,ku;kyo,ronFulouB;anj1Yl10miso2Era25; haGssaloni0U;gucigalpa,hr2Jl av0I;i0llinn,mpe26ngi04rtu;chu1Xn2HpQ;a3e2h1kopje,t0ydney;ockholm,uttga0Y;angh1Aenzh1S;o0HvW;int peters0Rlz0Rn 0ppo1A;jo0salvadV;se;eykjavik,i1o0;me,t23;ga,o de janei15;to;a8e6h5i4o2r0ueb1Oyongya1L;a0etor22;gue;rt0zn22; elizabe3o;ls1Erae22;iladelph1Xnom pe07oenix;r0tah tik17;th;lerJr0tr0Y;is;dessa,s0ttawa;a1Flo;a2ew 0is;delTtaip0york;ei;goya,nt0Spl0S;a5e4i3o1u0;mb0Jni0H;nt0scH;evideo,real;n01skolc;dellín,lbour0Q;drid,l5n3r0;ib1se0;ille;or;chest0dalay,i0X;er;mo;a4i1o0vAy00;ndZs angel0D;ege,ma0nz,sbYverpo1;!ss0;ol; pla0Gusan0D;a5hark4i3laipeda,o1rak0uala lump2;ow;be,pavog0sice;ur;ev,ng8;iv;b3mpa0Indy,ohsiu0Fra0un01;c0j;hi;ncheLstanb0̇zmir;ul;a5e3o0; chi mi1ms,u0;stH;nh;lsin0rakliF;ki;ifa,m0noi,va08;bu0QiltC;dan3en2hent,iza,othen1raz,ua0;dalaj0Engzhou;bu0N;eSoa;sk;es,rankfu0;rt;dmont4indhovT;a1ha00oha,u0;blQrb0Dshanbe;e0kar,masc0EugavpiI;gu,je0;on;a6ebu,h2o0raioIuriti00;lo0nstanIpenhagMrk;gEmbo;enn2i1ristchur0;ch;ang m0cago,ttagoK;ai;i0lgary,pe town,rac4;ro;aHeBirminghWogoAr5u0;char3dap3enos air2r0sZ;g0sa;as;es;est;a2isba1usse0;ls;ne;silPtisla0;va;ta;i3lgrade,r0;g1l0n;in;en;ji0rut;ng;ku,n3r0sel;celo1ranquil0;la;na;g1ja lu0;ka;alo0kok;re;aBb9hmedabad,l7m4n2qa1sh0thens,uckland;dod,gabat;ba;k0twerp;ara;m5s0;terd0;am;exandr0maty;ia;idj0u dhabi;an;lbo1rh0;us;rg","Currency":"$,aud,bPcNdIeurHfGgbp,hkd,inr,jpy,kElDp8r7s3usd,x2y1z0¢,£,¥,ден,лв,руб,฿,₡,₨,€,₭,﷼;lotyPł;en,uanO;af,of;h0terlings;e0il4;k0q0;elJ;iel,oubleIp,upeeI;e1ound ster0;lingG;n0soF;ceEn0;ies,y;empi6i6;n,r0wanzaByatB;!onaAw;ori7ranc9t;!o8;en3i2kk,o0;b0ll2;ra5;me4n0rham4;ar3;ad,e0ny;nt1;aht,itcoin0;!s","LastName":"0:31;1:2D;2:2Z;3:28;4:2P;a30b2Sc2Fd27e23f1Yg1Th1Ji1Ej19k12l0Vm0Jn0Fo0Cp02rVsKtFvDwBxAy7zh5;a5ou,u;ng,o;a5eun2Loshi1Eun;ma5ng;da,guc1Smo1Zsh1UzaO;iao,u;a5eb1o4right,u;gn1lk1ng,tanabe;a5ivaldi;ssilj2Xzqu0;a8h7i29o6r5sui,urn1;an,ynisH;lstoy,rr2Lth;at1Nomps2;kah0Rnaka;aCchBeAhimizu,i9mi8o7t6u5zabo;ar0lliv23zuC;al1Nein1;sa,u4;rn3th;lva,mmo1Yngh;mjon3rrano;midt,neid1ulz;ito,n6sa5to;ki;ch0dJz;amAeag1Ti8o6u5;bio,iz,sC;b5dri1HgGj0Pme1XsYux;erts,ins2;c5ve0C;ci,hards2;ir0os;aCe9h7ic5ow1T;as5hl1;so;a5illips;m,n1M;ders1Set6r0t5;e0Kr3;ers;h1Trk1t5vl3;el,te0H;baAg08liveiZr5;t5w1I;ega,iz;a5eils2guy1Kix2owak,ym19;gy,ka5var1E;ji5muU;ma;aDeBiAo7u5;ll1n5ssolini,ñ5;oz;lina,oIr5zart;al1Fe5;au,no;hhail3ll1;rci1s5y1;si;eTmmad3r5tsu04;in5tin0;!o;aAe7i5op0uo;!n5u;coln,dholm;e,fe5n0B;bv5v5;re;rs0Yu;aAennedy,imu9le07o7u6wo5;k,n;mar,znets3;bay5vacs;asV;ra;hn,rl8to,ur,zl3;ansse0Een8ha4imen0o5u4;h5n0Su4;an5ns2;ss2;ki0As0M;glesi8ke7noue,shik6to,vano5;u,v;awa;da;as;aAe8it7o6u5;!a4b1gh0Iynh;a4ffmann,rvat;chcock,l1;mingway,nde5rJ;rs2;ns0ArrOs6y5;asBes;an3hi5;moG;a7il,o6rub1u5;o,tierr0;m0nzal0;nd5o,rcia;hi;er8is7lor04o6uj5;ita;st1urni1;ch1;nand0;d7inste6sposi5vaK;to;in;is2wards;aviBeAi8omin7u5;bo5rand;is;gu0;az,mitr3;ov;lgado,vi;es,s;aEhAlarke,o5;hJl6op1x;er;em6li5;ns;an;an7e6iu,o5ristensEu4we;i,ng,u4w,y;!n,on5u4;!g;mpb6st5;ro;ell;a9e6ha4lanco,oyko,rooks,yrne;ng;ck6ethov5nnett;en;er,ham;ch,h6iley,rn5;es;k,ng;dams,l8nd5;ers5r9;en,on,s2;on;eks6iy7var0;ez;ej5;ev","Duration":"centur4d2hour3m0seconds,week3year3;i0onth2;llisecond1nute1;ay0ecade0;!s;ies,y","Date":"fri4mon4s2t1wed0;!nesd4;hurs2ues2;at0un1;!urd1;!d0;ay0;!s","Holiday":"0:1P;1:1Q;a1Fb1Bc12d0Ye0Of0Kg0Hh0Di09june07kwanzaa,l04m00nYoVpRrPsEt8v6w4xm03y2;om 2ule;hasho16kippur;hit2int0Xomens equalit7; 0Ss0T;aGe2ictor1E;r1Bteran0;-1ax 1h6isha bav,rinityNu2; b3rke2;y 1;ish2she2;vat;a0Ye prophets birth1;a6eptember15h4imchat tor0Vt 3u2;kk4mmer U;a9p8s7valentines day ;avu2mini atzeret;ot;int 2mhain;a5p4s3va2;lentine0;tephen0;atrick0;ndrew0;amadan,ememberanc0Yos2;a park0h hashana;a3entecost,reside0Zur2;im,ple heart 1;lm2ssovE; s04;rthodox 2stara;christma0easter2goOhoJn0C;! m07;ational 2ew years09;freedom 1nurse0;a2emorial 1lHoOuharram;bMr2undy thurs1;ch0Hdi gr2tin luther k0B;as;a2itRughnassadh;bour 1g baom2ilat al-qadr;er; 2teenth;soliU;d aJmbolc,n2sra and miraj;augurGd2;ependen2igenous people0;c0Bt0;a3o2;ly satur1;lloween,nukkUrvey mil2;k 1;o3r2;ito de dolores,oundhoW;odW;a4east of 2;our lady of guadalupe,the immaculate concepti2;on;ther0;aster8id 3lectYmancip2piphany;atX;al-3u2;l-f3;ad3f2;itr;ha;! 2;m8s2;un1;ay of the dead,ecemb3i2;a de muertos,eciseis de septiembre,wali;er sol2;stice;anad8h4inco de mayo,o3yber m2;on1;lumbu0mmonwealth 1rpus christi;anuk4inese n3ristmas2;! N;ew year;ah;a 1ian tha2;nksgiving;astillCeltaine,lack4ox2;in2;g 1; fri1;dvent,ll 9pril fools,rmistic8s6u2;stral4tum2;nal2; equinox;ia 1;cens2h wednes1sumption of mary;ion 1;e 1;hallows 6s2;ai2oul0t0;nt0;s 1;day;eve","Person":"aQbOcKdHemeril lagasse,faGgDhBjk rowling,kAlebron james,m7oprah winfrNparis hiltIr4s3t2uncle,v0womF;a0irgin maB;lentino rossi,n go3;iger woods,yra banks;addam hussaHcarlett johanssEistOlobodan milosevic,omeone,tepA;ay romano,eese witherspoDo1ush limbau0;gh;d stewart,naldinho;an,essiaen,itt romnFo0ubarek;m0thJ;!my;anye west,iefer sutherland,obe bryaG;alle ber0ulk hog3;ry;entlem1irl,rand0uy;fa1mo1;an;thB;ad1enzel washingt0ick wolf,ude;on;!dy;ar1ous0;in;dinal wols0son palm5;ey;arack obama,oy,ro0;!th2;dolf hitl1shton kutch1u0;nt;er","Conjunction":"aDb9cuz,for,how7in caCno6o5p3th1vers4wh0yet;eth7ile;eref9o0;!uC;l0rovided that;us;r,therwi6; matt1r;!ev0;er;e0ut;cau1f0;ore;se;lthou1nd,s 0;far as,if;gh","PhrasalVerb":"0:71;1:6P;2:7D;3:73;4:6I;5:7G;6:75;7:6O;8:6B;9:6C;A:5H;B:70;C:6Z;a7Gb62c5Cd59e57f45g3Nh37iron0j33k2Yl2Km2Bn29o27p1Pr1Es09tQuOvacuum 1wGyammerCzD;eroAip EonD;e0k0;by,up;aJeGhFiEorDrit52;d 1k2Q;mp0n49pe0r8s8;eel Bip 7K;aEiD;gh 06rd0;n Br 3C;it 5Jk8lk6rm 0Qsh 73t66v4O;rgeCsD;e 9herA;aRePhNiJoHrFuDype 0N;ckArn D;d2in,o3Fup;ade YiDot0y 32;ckle67p 79;ne66p Ds4C;d2o6Kup;ck FdEe Dgh5Sme0p o0Dre0;aw3ba4d2in,up;e5Jy 1;by,o6U;ink Drow 5U;ba4ov7up;aDe 4Hll4N;m 1r W;ckCke Elk D;ov7u4N;aDba4d2in,o30up;ba4ft7p4Sw3;a0Gc0Fe09h05i02lYmXnWoVpSquare RtJuHwD;earFiD;ngEtch D;aw3ba4o6O; by;ck Dit 1m 1ss0;in,up;aIe0RiHoFrD;aigh1LiD;ke 5Xn2X;p Drm1O;by,in,o6A;r 1tc3H;c2Xmp0nd Dr6Gve6y 1;ba4d2up;d2o66up;ar2Uell0ill4TlErDurC;ingCuc8;a32it 3T;be4Brt0;ap 4Dow B;ash 4Yoke0;eep EiDow 9;c3Mp 1;in,oD;ff,v7;gn Eng2Yt Dz8;d2o5up;in,o5up;aFoDu4E;ot Dut0w 5W;aw3ba4f36o5Q;c2EdeAk4Rve6;e Hll0nd GtD; Dtl42;d2in,o5upD;!on;aw3ba4d2in,o1Xup;o5to;al4Kout0rap4K;il6v8;at0eKiJoGuD;b 4Dle0n Dstl8;aDba4d2in52o3Ft2Zu3D;c1Ww3;ot EuD;g2Jnd6;a1Wf2Qo5;ng 4Np6;aDel6inAnt0;c4Xd D;o2Su0C;aQePiOlMoKrHsyc29uD;ll Ft D;aDba4d2in,o1Gt33up;p38w3;ap37d2in,o5t31up;attleCess EiGoD;p 1;ah1Gon;iDp 52re3Lur44wer 52;nt0;ay3YuD;gAmp 9;ck 52g0leCn 9p3V;el 46ncilA;c3Oir 2Hn0ss FtEy D;ba4o4Q; d2c1X;aw3ba4o11;pDw3J;e3It B;arrow3Serd0oD;d6te3R;aJeHiGoEuD;ddl8ll36;c16p 1uth6ve D;al3Ad2in,o5up;ss0x 1;asur8lt 9ss D;a19up;ke Dn 9r2Zs1Kx0;do,o3Xup;aOeMiHoDuck0;a16c36g 0AoDse0;k Dse34;aft7ba4d2forw2Ain3Vov7uD;nd7p;e GghtFnEsDv1T;ten 4D;e 1k 1; 1e2Y;ar43d2;av1Ht 2YvelD; o3L;p 1sh DtchCugh6y1U;in3Lo5;eEick6nock D;d2o3H;eDyA;l2Hp D;aw3ba4d2fSin,o05to,up;aFoEuD;ic8mpA;ke2St2W;c31zz 1;aPeKiHoEuD;nker2Ts0U;lDneArse2O;d De 1;ba4d2oZup;de Et D;ba4on,up;aw3o5;aDlp0;d Fr Dt 1;fDof;rom;in,oO;cZm 1nDve it;d Dg 27kerF;d2in,o5;aReLive Jloss1VoFrEunD; f0M;in39ow 23; Dof 0U;aEb17it,oDr35t0Ou12;ff,n,v7;bo5ft7hJw3;aw3ba4d2in,oDup,w3;ff,n,ut;a17ek0t D;aEb11d2oDr2Zup;ff,n,ut,v7;cEhDl1Pr2Xt,w3;ead;ross;d aEnD;g 1;bo5;a08e01iRlNoJrFuD;cDel 1;k 1;eEighten DownCy 1;aw3o2L;eDshe1G; 1z8;lFol D;aDwi19;bo5r2I;d 9;aEeDip0;sh0;g 9ke0mDrD;e 2K;gLlJnHrFsEzzD;le0;h 2H;e Dm 1;aw3ba4up;d0isD;h 1;e Dl 11;aw3fI;ht ba4ure0;eInEsD;s 1;cFd D;fDo1X;or;e B;dQl 1;cHll Drm0t0O;apYbFd2in,oEtD;hrough;ff,ut,v7;a4ehi1S;e E;at0dge0nd Dy8;o1Mup;o09rD;ess 9op D;aw3bNin,o15;aShPlean 9oDross But 0T;me FoEuntD; o1M;k 1l6;aJbIforGin,oFtEuD;nd7;ogeth7;ut,v7;th,wD;ard;a4y;pDr19w3;art;eDipA;ck BeD;r 1;lJncel0rGsFtch EveA; in;o16up;h Bt6;ry EvD;e V;aw3o12;l Dm02;aDba4d2o10up;r0Vw3;a0He08l01oSrHuD;bbleFcklTilZlEndlTrn 05tDy 10zz6;t B;k 9; ov7;anMeaKiDush6;ghHng D;aEba4d2forDin,o5up;th;bo5lDr0Lw3;ong;teD;n 1;k D;d2in,o5up;ch0;arKgJil 9n8oGssFttlEunce Dx B;aw3ba4;e 9; ar0B;k Bt 1;e 1;d2up; d2;d 1;aIeed0oDurt0;cFw D;aw3ba4d2o5up;ck;k D;in,oK;ck0nk0st6; oJaGef 1nd D;d2ov7up;er;up;r0t D;d2in,oDup;ff,ut;ff,nD;to;ck Jil0nFrgEsD;h B;ainCe B;g BkC; on;in,o5; o5;aw3d2o5up;ay;cMdIsk Fuction6; oD;ff;arDo5;ouD;nd;d D;d2oDup;ff,n;own;t D;o5up;ut","Modal":"c5m4ought3sh1w0;ill,o5;a0o4;ll,nt;! to;ay,ight,ust;an,o0;uld","PerfectTense":"have taken","pluPerfectTense":"had taken","FuturePerfect":"will have taken"}
